% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{add_features_metadata}
\alias{add_features_metadata}
\title{Adding data-frame to features metadata}
\usage{
add_features_metadata(mo_data, df)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{df}{A tibble or data-frame of features information, with at least
columns \code{feature_id} (giving the feature IDs) and \code{dataset} (giving the
name of the dataset to which the features belong), and one row per feature
ID. Can contain info about features from different datasets.}
}
\value{
A MultiDataSet object, with info from \code{df} adding to its
corresponding features metadata.
}
\description{
Adds information from a data-frame to the features metadata of a MultiDataSet
object.
}
