% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_classes_eSet.R
\name{add_metabo,MultiDataSet,MetabolomeSet-method}
\alias{add_metabo,MultiDataSet,MetabolomeSet-method}
\title{Adds a MetabolomeSet to a MultiDataSet object.}
\usage{
\S4method{add_metabo}{MultiDataSet,MetabolomeSet}(object, met_set, warnings = TRUE, ...)
}
\arguments{
\item{object}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{met_set}{A \linkS4class{MetabolomeSet} object.}

\item{warnings}{Logical, should warnings be displayed? Default is \code{TRUE}.}

\item{...}{Further arguments passed to the \code{\link[MultiDataSet:add_eset]{MultiDataSet::add_eset()}}
function.}
}
\value{
A new \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object with a slot filled.
}
\description{
Adds a MetabolomeSet to a MultiDataSet object.
}
