% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_omics_sets.R
\name{add_omics_set}
\alias{add_omics_set}
\title{Adds an omics set to a MultiDataSet object}
\usage{
add_omics_set(mo_data, omics_set, ds_name, ...)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{omics_set}{A \link[Biobase:class.eSet]{Biobase::eSet} object, created via
\code{\link[=create_omics_set]{create_omics_set()}}. Currently accepted objects: \link[Biobase:class.SnpSet]{Biobase::SnpSet},
\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet}, \linkS4class{MetabolomeSet},
\linkS4class{PhenotypeSet}.}

\item{ds_name}{Character, name of the dataset (will be used as suffix for the
name of the dataset in the resulting MultiDataSet object).}

\item{...}{Further arguments passed to \verb{[MultiDataSet::add_snps()], [MultiDataSet::add_rnaseq()], [add_metabo()] or[add_pheno()] (depending on }omics_set` class).}
}
\value{
A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object, the \code{mo_data} with
\code{omics_set} as an additional dataset.
}
\description{
Adds a omics set to an existing MultiDataSet object.
}
\examples{
\dontrun{
add_omics_set(mo_data, omics_set, "exp1")
}
}
