% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_classes_eSet.R
\name{add_pheno}
\alias{add_pheno}
\title{Method to add a PhenotypeSet to a MultiDataSet object.}
\usage{
add_pheno(object, pheno_set, warnings = TRUE, ...)
}
\arguments{
\item{object}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{pheno_set}{A \linkS4class{PhenotypeSet} object.}

\item{warnings}{Logical, should warnings be displayed? Default is \code{TRUE}.}

\item{...}{Further arguments passed to the \code{\link[MultiDataSet:add_eset]{MultiDataSet::add_eset()}}
function.}
}
\value{
A new \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object with a slot filled.
}
\description{
Method to add a PhenotypeSet to a MultiDataSet object.
}
