% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_running_time.R
\name{aggr_patterns_fct}
\alias{aggr_patterns_fct}
\title{Aggregate regexp patterns for a search.}
\usage{
aggr_patterns_fct(x)
}
\arguments{
\item{x}{Character vector, the regex patterns.}
}
\value{
Character, aggregated regex patterns.
}
\description{
Aggregate regexp patterns for a search.
}
