% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_evaluation.R
\name{check_feature_sets}
\alias{check_feature_sets}
\title{Checks features assignment to sets}
\usage{
check_feature_sets(feature_sets, mo_data, datasets = names(mo_data))
}
\arguments{
\item{feature_sets}{Named list, where each element corresponds to a feature set,
and contains a vector of features ID of all features belonging to that set.}

\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{datasets}{Character vector, the names of the datasets for which features
assignment should be checked. By default, all datasets in \code{mo_data} are considered.}
}
\value{
A tibble giving for each dataset the number and fraction of features
that are assigned to at least one feature set. The \code{message} column is meant to
facilitate reporting.
}
\description{
Checks proportion of features from a multi-omics dataset that are assigned to
feature sets.
}
