% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{check_input_multidataset}
\alias{check_input_multidataset}
\title{Check a MultiDataSet input}
\usage{
check_input_multidataset(x, datasets = NULL)
}
\arguments{
\item{x}{A input object that hopefully is  a \code{MultiDataSet} object.}

\item{datasets}{Character vector of dataset names that should be in \code{x}. If
\code{NULL} (default value), dataset names will not be checked}
}
\value{
The MultiDataSet object restricted to the datasets required (if
\code{datasets} is not \code{NULL}.
}
\description{
Checks a MultiDataSet object provided as an input. In particular, checks that
\enumerate{
\item the input object is a \code{MultiDataSet} object, 2) the datasets stored match
the datasets named provided (if any). Will restrict the MultiDataSet object
to only necessary datasets.
}
}
