% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{check_is_multidataset}
\alias{check_is_multidataset}
\title{Checks whether object is MultiDataSet}
\usage{
check_is_multidataset(x)
}
\arguments{
\item{x}{An object.}
}
\value{
nothing.
}
\description{
Checks whether an input object is a \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet}
object.
}
