% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_evaluation.R
\name{compute_samples_silhouette}
\alias{compute_samples_silhouette}
\title{Computes samples silhouette score from method output}
\usage{
compute_samples_silhouette(
  method_output,
  mo_data,
  group_col,
  latent_dimensions = NULL,
  distance_metric = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski")
)
}
\arguments{
\item{method_output}{method_output Integration method output generated via the
\code{get_output()} function.}

\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{group_col}{Character, name of the column in one of the samples metadata
table from \code{mo_data} containing the samples grouping to be used.}

\item{latent_dimensions}{Character vector, the latent dimensions to use for
computing distance between samples. If \code{NULL} (default value), all latent
dimensions will be used.}

\item{distance_metric}{Character, name of the metric to use when computing
the distance between samples from their coordinates in the latent dimensions.
This will be passed to the \code{\link[stats:dist]{stats::dist()}} function. Options include: \code{"euclidean"}
(default value), \code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.}
}
\value{
A list with two elements:
\itemize{
\item \code{samples_silhouette}: a tibble giving for each sample (\code{sample_id} column)
the group to which it belongs (\code{group} column), its closest (other) group in the
space spanned by the latent dimensions (\code{neighbour_group}), and its silhouette
width (\code{silhouette_width} column).
\item \code{groups_average_silhouette}: a tibble giving for each samples group (\code{group}
column) its average silhouette width (\code{group_average_width} column).
}
}
\description{
Calculates the samples silhouette width from the results of a dimension
reduction method, according to some samples grouping from the samples
metadata of a \code{MultiDataSet} object.
}
\details{
The samples silhouette width and groups average width are calculated using
the \code{\link[cluster:silhouette]{cluster::silhouette()}} function.
}
