% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_comparison.R
\name{consensus_importance_metric}
\alias{consensus_importance_metric}
\title{Calculate features importance score}
\usage{
consensus_importance_metric(x, metric, offset = 0)
}
\arguments{
\item{x}{Numeric vector of importance values (between 0 and 1).}

\item{metric}{Character, one of the metrics to use to compute the consensus score.
Can be one of \code{'min'}, \code{'max'}, \code{'average'}, \code{'product'}, \code{'l2'} (for L2-norm),
\code{'geometric'} (for geometric mean) or \code{'harmonic'} (for harmonic mean).}

\item{offset}{Numeric (strictly positive), used to replace zero values to compute the geometric and
harmonic mean. If \code{0} (default value), the zero values will be ignored when
calculating the geometric and harmonic mean. Accepting vector of values to facilitate use
whithin \code{dplyr::summarise()}.}
}
\value{
A numeric value, the importance consensus score.
}
\description{
Given a vector of features importance (between 1 and 0), returns a consensus importance
score.
}
