% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_report.R
\name{create_report}
\alias{create_report}
\title{Creates Rmd report from template}
\usage{
create_report(
  file,
  add_sections = c("spls", "so2pls", "mofa", "diablo", "comparison"),
  overwrite = FALSE,
  target_project = Sys.getenv("TAR_PROJECT", "main"),
  use_quarto = FALSE
)
}
\arguments{
\item{file}{Name (and path) of the file to be created. Should end with \code{.Rmd}
if \code{use_quarto} is \code{FALSE}, or \code{.qmd} if \code{use_quarto} is \code{TRUE}.}

\item{add_sections}{Character vector, names of the sections to include in the
report. Possible values are 'spls', 'so2pls', 'mofa', 'diablo' and 'comparison'.
By default, all sections are included.}

\item{overwrite}{Logical, should existing file be overwritten?}

\item{target_project}{Character, name of the current targets project (i.e. value to be used for
the \code{TAR_PROJECT} environment variable). If none provided, will be read from the \code{TAR_PROJECT}
environment variable or set to \code{"main"} if the former is not set.}

\item{use_quarto}{Boolean, whether to use a Quarto report. Default value is \code{FALSE}, i.e.
uses a Rmarkdown report.}
}
\value{
Invisible character, the path and name of the generated \code{.Rmd} (or \code{.qmd}) file.
}
\description{
Creates a Rmarkdown report to present the results from the integration analysis.
This function only creates a \code{.Rmd} file and does not knit the document.
}
