% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_report.R
\name{create_targets_pipeline}
\alias{create_targets_pipeline}
\title{Creates a target script file from template}
\usage{
create_targets_pipeline(file = "_targets.R", overwrite = FALSE)
}
\arguments{
\item{file}{Name (and path) of the file to be created. Should end with \code{.R}.
Default value (recommended) is \code{"_targets.R"} (in the current directory).}

\item{overwrite}{Logical, should existing file be overwritten?}
}
\value{
The file name (invisibly).
}
\description{
Creates a target script file form a template multi-omics integration pipeline.
This function only creates the script file but does not execute it.
}
