% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_get_pairwise_pls_corr}
\alias{diablo_get_pairwise_pls_corr}
\title{Get pairwise correlations from PLS run}
\usage{
diablo_get_pairwise_pls_corr(pairwise_pls_result)
}
\arguments{
\item{pairwise_pls_result}{List containing the results from the pairwise PLS runs, computed by the \code{\link{run_pairwise_pls}} function.}
}
\value{
A matrix of correlation coefficients between datasets.
}
\description{
Computes the correlation matrix between datasets based on the first component of a PLS run.
}
\details{
The correlation coefficient between two datasets is computed as the correlation coefficient between the first component of each
dataset obtained from a Projection to Latent Structure (PLS) run, from the \code{mixOmics} package.
}
\examples{
\dontrun{
## get mixomics input from MultiDataSet object
mixomics_data <- get_input_mixomics_supervised(mo_set, "outcome_group")

## Get the list of dataset names
datasets <- setdiff(names(mixomics_data), "Y")

## Get all possible pairwise combinations of dataset names
ds_pairs <- utils::combn(datasets, 2)

## run PLS for each pair of datasets
pls_res_list <- lapply(1:ncol(ds_pairs), function(i) {
  run_pairwise_pls(mo_set, ds_pairs[, i])
})

## extract the pairwise correlation matrix
diablo_get_pairwise_pls_corr(pls_res_list)
}
}
