% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_get_params}
\alias{diablo_get_params}
\title{Get parameters from DIABLO run}
\usage{
diablo_get_params(diablo_res)
}
\arguments{
\item{diablo_res}{The output from \code{\link[mixOmics]{block.splsda}} or \code{\link{diablo_run}}.}
}
\value{
A tibble.
}
\description{
Extracts the \code{ncomp} and \code{keepX} parameters from a DIABLO run and format them into a table.
}
