% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_get_wa_coord}
\alias{diablo_get_wa_coord}
\title{Get weighted average coordinates}
\usage{
diablo_get_wa_coord(diablo_res)
}
\arguments{
\item{diablo_res}{The output from \code{\link[mixOmics]{block.splsda}} or \code{\link{diablo_run}}.}
}
\value{
A matrix with one row per sample and one column per latent component.
}
\description{
Computes the samples coordinates in the weighted average latent components space from a DIABLO result object.
}
