% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_plot}
\alias{diablo_plot}
\title{Plots DIABLO output}
\usage{
diablo_plot(
  diablo_res,
  ncomp = 1,
  legend = TRUE,
  legend.ncol,
  col.per.group = NULL
)
}
\arguments{
\item{diablo_res}{The output from \code{\link[mixOmics]{block.splsda}} or \code{\link{diablo_run}}.}

\item{ncomp}{Integer, the latent component to plot.}

\item{legend}{Logical, should the legend be added to the plot? Default value is \code{TRUE}.}

\item{legend.ncol}{Integer, the number of columns for the legend. If none specified, will be calculated as
\code{min(5, nlevels(diablo_res$Y))}.}

\item{col.per.group}{Named character vector, provides the colours to use for each phenotypic group. Names
must match the levels of \code{diablo_res$Y}.}
}
\value{
None.
}
\description{
Displays the samples coordinates for a given latent component across the datasets.
This is a copy of the \code{\link[mixOmics]{plotDiablo}} function, the only difference is that the
margins have been increased to accomodate for a title.
}
