% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_plot_circos}
\alias{diablo_plot_circos}
\title{Plots DIABLO circos plot}
\usage{
diablo_plot_circos(diablo_res, mo_data, label_cols, truncate = NULL, ...)
}
\arguments{
\item{diablo_res}{The output from \code{\link[mixOmics]{block.splsda}} or \code{\link{diablo_run}}.}

\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{label_cols}{Character or named list of character, giving for each
dataset the name of the column in the corresponding features metadata to
use as label. If one value, will be used for all datasets. If list, the
names must correspond to the names of the datasets in \code{mo_data}. Default
value is \code{feature_id}, i.e. by default the ID of the features will be used
as label.}

\item{truncate}{Integer, width to which the labels should be truncated (to
avoid issues with very long labels in plots). If \code{NULL} (default value), no
truncation will be performed.}

\item{...}{Additional arguments passed to \code{\link[mixOmics]{circosPlot}}.}
}
\description{
Displays the DIABLO circos plot, but uses available feature metadata
to display feature names.
}
