% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_plot_tune}
\alias{diablo_plot_tune}
\title{Plots DIABLO tune results}
\usage{
diablo_plot_tune(tune_res)
}
\arguments{
\item{tune_res}{The cross-validation results, computed with \code{\link{diablo_tune}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Displays the error rate of a DIABLO run cross-validation to estimate the optimal number of features to retain from
each dataset (\code{keepX}).
}
