% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_plot_var}
\alias{diablo_plot_var}
\title{Plots DIABLO features correlation circle}
\usage{
diablo_plot_var(
  diablo_res,
  mo_data,
  label_cols = "feature_id",
  truncate = NULL,
  ...
)
}
\arguments{
\item{diablo_res}{The output from \code{\link[mixOmics]{block.splsda}} or \code{\link{diablo_run}}.}

\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{label_cols}{Character or named list of character, giving for each
dataset the name of the column in the corresponding features metadata to
use as label. If one value, will be used for all datasets. If list, the
names must correspond to the names of the datasets in \code{mo_data}. Default
value is \code{feature_id}, i.e. by default the ID of the features will be used
as label.}

\item{truncate}{Integer, width to which the labels should be truncated (to
avoid issues with very long labels in plots). If \code{NULL} (default value), no
truncation will be performed.}

\item{...}{Additional arguments passed to \code{\link[mixOmics]{plotVar}}.}
}
\value{
A plot (see \code{\link[mixOmics]{plotVar}}).
}
\description{
Displays the DIABLO correlation circle plot, but uses available feature metadata
to display feature names.
}
\examples{
\dontrun{
# Use the default features ID for the plot
diablo_plot_var(
  diablo_final_run,
  mo_data,
  "feature_id",
  overlap = FALSE,
  cex = rep(2, 3),
  comp = 1:2
)

# Using a different column from the feature metadata of each omics dataset
diablo_plot_var(
  diablo_final_run,
  mo_presel_supervised,
  c(
    "snps" = "feature_id",
    "rnaseq" = "gene_name",
    "metabolome" = "comp_name"
  ),
  overlap = FALSE,
  cex = rep(2, 3),
  comp = 1:2
)
}
}
