% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_predefined_design_matrix}
\alias{diablo_predefined_design_matrix}
\title{Generate a design matrix for DIABLO}
\usage{
diablo_predefined_design_matrix(
  datasets_name,
  design_matrix = c("null", "weighted_full", "full")
)
}
\arguments{
\item{datasets_name}{Character vector, the names of the datasets to
integrate. Should include the value \code{"Y"} to represent the samples outcome
groups.}

\item{design_matrix}{Character, the type of design matrix to generate. Should
be one of \code{"null"}, \code{"weighted_full"} or \code{"full"}.}
}
\value{
A matrix with as many rows and columns as the length of
\code{datasets_name}, and filled with either 0 (\code{design_matrix = "null"}), 0.1
(\code{design_matrix = "weighted_full"}) or 1 (\code{design_matrix = "full"}). Values
in the diagonal are set to 0, and values in the \code{"Y"} row and columns are
set to 1.
}
\description{
Generates a design matrix for DIABLO, following a predesigned pattern as
recommended by the mixOmics authors.
}
