% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_run}
\alias{diablo_run}
\title{Runs DIABLO algorithm}
\usage{
diablo_run(mixomics_data, design_matrix, ...)
}
\arguments{
\item{mixomics_data}{A \code{mixOmics} input object created with \code{\link{get_input_mixomics_supervised}}.}

\item{design_matrix}{Either numeric matrix created through \code{\link{diablo_generate_design_matrix}}, or character
(accepted values are \code{'null'}, \code{'weighted_full'}, \code{'full'}). See Details.}

\item{...}{Arguments to be passed to the \code{\link[mixOmics]{block.splsda}} function.}
}
\value{
An object of class \code{block.splsda} (if \code{keepX} argument was provided) or \code{block.splsda}
(if it was not), see \code{\link[mixOmics:block.splsda]{mixOmics::block.splsda()}} and \code{\link[mixOmics:block.plsda]{mixOmics::block.plsda()}}.
}
\description{
Runs the DIABLO algorithm (\code{\link[mixOmics]{block.splsda}}) from the \code{mixOmics} package.
}
\details{
The \code{design_matrix} argument can either be a custom design matrix (for example as constructed via the
\code{\link{diablo_generate_design_matrix}} function); or a character indicating the type of design matrix
to generate. Possible values include:
\itemize{
\item \code{'null'}: Off-diagonal elements of the design matrix are set to 0;
\item \code{'weighted_full'}: Off-diagonal elements of the design matrix are set to 0.1;
\item \code{'full'}: Off-diagonal elements of the design matrix are set to 1.
}
}
