% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{diablo_table_optim_keepX}
\alias{diablo_table_optim_keepX}
\title{Formatted table with optimal keepX values}
\usage{
diablo_table_optim_keepX(tune_res)
}
\arguments{
\item{tune_res}{The cross-validation results, computed with \code{\link{diablo_tune}}.}
}
\value{
A tibble with a Dataset column, one column for each latent component and a Total column giving the number of
features to retain from the corresponding dataset across all latent components.
}
\description{
Produces a nicely formatted table with the optimal number of features to select from each dataset for a
DIABLO run according to the results of a cross-validation analysis. Used mostly for producing a nice report.
}
