% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_features_labels_toplot}
\alias{.add_features_labels_toplot}
\title{Adds features label to data-frame}
\usage{
.add_features_labels_toplot(toplot, label_cols, mo_data, truncate = NULL)
}
\arguments{
\item{toplot}{The data-frame to which the labels should be added.}

\item{label_cols}{Character or named list of character, giving for each
dataset the name of the column in the corresponding features metadata to
use as label. If one value, will be used for all datasets. If list, the
names must correspond to the names of the datasets in \code{mo_data}. If a
dataset is missing from the list or no value is provided, feature IDs will
be used as labels. Alternatively, use \code{feature_id} to get the feature IDs
as labels.}

\item{mo_data}{A \code{MultiDataSet} object. Only used if \code{label_cols} is not
\code{NULL}.}

\item{truncate}{Integer, width to which the labels should be truncated (to
avoid issues with very long labels in plots). If \code{NULL} (default value), no
truncation will be performed.}
}
\value{
the \code{toplot} data-frame with an additional column \code{label}.
}
\description{
Adds the features label to a data-frame for plotting. Can be extracted from
the features metadata of a \code{MultiDataSet} object; otherwise use the feature
IDs as label. If some labels are missing, feature IDs will be used instead.
}
