% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_input_var_smetadata_common}
\alias{.check_input_var_smetadata_common}
\title{Check that variable names corresponds to columns in samples metadata}
\usage{
.check_input_var_smetadata_common(x, mo_data)
}
\arguments{
\item{x}{Character, name of the column from the samples metadata.}

\item{mo_data}{A \code{MultiDataSet} object containing samples information for the
datasets. Should be checked with \code{.check_input_multidataset()}.}
}
\value{
Nothing. Will throw an error if need be.
}
\description{
Checks whether a variable name corresponds to a column in the samples
metadata of the corresponding dataset. If one value is provided, will be used
for all datasets.
}
