% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{.check_names_output_list}
\alias{.check_names_output_list}
\title{Check names of output list}
\usage{
.check_names_output_list(output_list)
}
\arguments{
\item{output_list}{List of integration methods output, each generated via one of the
\verb{get_output_*()} function.}
}
\value{
output_list (but named if it wasn't).
}
\description{
Checks that the names of a list of outputs from several integration methods
are unique. If not named, the name of the method will be used as name.
}
