% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{.clean_seed}
\alias{.clean_seed}
\title{Clean seed for sPLS-DA preselection factory}
\usage{
.clean_seed(x, ds)
}
\arguments{
\item{x}{Integer vector of seeds to use.}

\item{ds}{Character vector of dataset names.}
}
\value{
\code{x} with names if it didn't have them or an error.
}
\description{
Checks seed input arguments for the \code{\link[=feature_preselection_splsda_factory]{feature_preselection_splsda_factory()}} function.
}
