% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{.filter_output_datasets}
\alias{.filter_output_datasets}
\title{Filter datasets}
\usage{
.filter_output_datasets(
  method_output,
  datasets,
  fixed_length = NULL,
  method_name = attr(method_output, "method")
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{datasets}{Character vector giving the datasets to retain in the features
weight table of the method's output.}

\item{fixed_length}{Integer, expected length of \code{datasets}.
If \code{NULL} (default value), the length of \code{datasets} will not be checked.}

\item{method_name}{Character, name of the method to use in the error message.}
}
\value{
Similar to \code{method_output}, but the features weight table has been filtered.
}
\description{
Filters datasets by name in the output of an integration method.
}
