% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{.filter_output_dimensions}
\alias{.filter_output_dimensions}
\title{Filter latent dimensions}
\usage{
.filter_output_dimensions(
  method_output,
  latent_dimensions,
  fixed_length = NULL,
  method_name = attr(method_output, "method")
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{latent_dimensions}{Character vector giving the latent dimensions to retain
in the method's output.}

\item{fixed_length}{Integer, expected length of \code{latent_dimensions}.
If \code{NULL} (default value), the length of \code{latent_dimensions} will not be checked.}

\item{method_name}{Character, name of the method to use in the error message.}
}
\value{
Similar to \code{method_output}, but the samples score table
or features weight table have been filtered.
}
\description{
Filters latent dimensions by name in the output of an integration method.
}
