% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_comparison.R
\name{.get_initials}
\alias{.get_initials}
\title{Get initials from sentence}
\usage{
.get_initials(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
A character vector, each element containing the initials of the words
in \code{x} in upper case plus any number, all pasted.
}
\description{
Extracts initials from a sentence as well as any number.
}
