% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_comparison.R
\name{.heatmap_samples_score_corr}
\alias{.heatmap_samples_score_corr}
\title{Heatmap of samples score correlation}
\usage{
.heatmap_samples_score_corr(output_list, hclust_fw)
}
\arguments{
\item{output_list}{List of integration methods output generated via the
\code{get_output()} function.}

\item{hclust_fw}{Dendrogram of latent dimensions according to the features
weight correlation (obtained with \code{.heatmap_features_weight_corr}).}
}
\value{
a \code{ComplexHeatmap::Heatmap} (upper triangle only).
}
\description{
Constructs a upper triangle heatmap of samples score correlation between
the latent dimensions constructed by several integration methods.
}
