% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_evaluation.R
\name{evaluate_feature_selection_table}
\alias{evaluate_feature_selection_table}
\title{Evaluate feature selection against features label}
\usage{
evaluate_feature_selection_table(
  method_output,
  mo_data,
  col_names,
  latent_dimensions = NULL
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{col_names}{Named character vector, giving for each dataset the
name of the column in the features metadata table that contains the features
label. If a dataset is not present in this vector, will be excluded from the
resulting table.}

\item{latent_dimensions}{Character vector, the latent dimensions to include
in the resulting table. If \code{NULL} (default value), all latent dimensions will
be represented.}
}
\value{
A tibble, with for each dataset and latent dimension the number of
selected and non-selected features per feature label.
}
\description{
Compares the selection of features with different feature labels (e.g.
result of a DE analysis) for each latent dimension.
}
