% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_pca.R
\name{get_complete_data}
\alias{get_complete_data}
\title{Get MultiDataSet object with imputed values}
\usage{
get_complete_data(mo_data, pca_result)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{pca_result}{A list in which each element is the result of a PCA run on
a different dataset, computed with the \code{\link[=run_pca]{run_pca()}} function.}
}
\value{
A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object, for which the assay of
each dataset is the imputed dataset.
}
\description{
Replace missing values with imputed values for each dataset of a MultiDataSet
object, based on the results of a Principal Component Analysis applied to the
corresponding dataset.
}
\details{
Uses the \code{\link[pcaMethods:completeObs-nniRes-method]{pcaMethods::completeObs()}} function to impute missing values.
}
