% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{get_datasets}
\alias{get_datasets}
\title{Get multi-omics measurement datasets}
\usage{
get_datasets(mo_data)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
A named list of matrices, each with features as rows and samples as
columns.
}
\description{
Returns the multi-omics datasets as a list of matrices from a MultiDataSet
object.
}
