% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{get_features}
\alias{get_features}
\title{Get feature IDs from MultiDataSet}
\usage{
get_features(mo_data)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
A named list, with one element per dataset, and each element is a
character vector of feature IDs.
}
\description{
Extract the list of feature IDs from each dataset in a MultiDataSet object.
}
