% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{get_features_metadata}
\alias{get_features_metadata}
\title{Get features metadata dataframes from MultiDataSet}
\usage{
get_features_metadata(mo_data)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
A named list of data-frames, one per dataset in the \code{mo_data}
object.
}
\description{
Extracts the features metadata dataframe (featureData field) from each
dataset for a MultiDataSet object.
}
