% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{get_filtered_dataset_variability}
\alias{get_filtered_dataset_variability}
\title{Get filtered MultiDataSet object based on variability measure}
\usage{
get_filtered_dataset_variability(mo_data, var_list)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{var_list}{A list with the result from the MAD or COV calculation for each dataset to be filtered, as returned by
the \code{\link{select_features_mad}} or \code{\link{select_features_cov}} function.}
}
\value{
A \code{\link[MultiDataSet]{MultiDataSet-class}} object.
}
\description{
Selects most highly variable features from omics datasets based on features'
variability (e.g. MAD or COV).
}
\examples{
\dontrun{
# Goal: keep 20\% of features in dataset1, and 50\% of features in dataset2
to_keep_prop <- c("dataset1" = 0.2, "dataset_2" = 0.5)

# 1) compute MAD values and select features for dataset1 and dataset2
mad_list <- lapply(names(to_keep_prop), function(i) {
  select_features_mad(mo_data, i, to_keep_prop[i])
})

# 2) Get the filtered dataset
mo_data_filtered <- get_filtered_dataset_variability(mo_data, mad_list)
}
}
