% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefisto.R
\name{get_input_mefisto}
\alias{get_input_mefisto}
\title{Generate MEFISTO input data}
\usage{
get_input_mefisto(
  mo_data,
  covariates,
  datasets = names(mo_data),
  groups = NULL,
  options_list = NULL,
  only_common_samples = FALSE
)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{covariates}{Character or character vector of length 2, the column name(s) in the samples metadata
data-frames to use as continuous covariates.}

\item{datasets}{Character vector, the names of the datasets from \code{mo_data} to include in the analysis.}

\item{groups}{Character, the column name in the samples metadata data-frames to use as groups
(use \code{\link{get_samples_metadata}} to view the samples metadata data-frame for each dataset).}

\item{options_list}{A named list. Should contain at most 4 elements, named 'data_options', 'model_options', 'training_options' and 'mefisto_options'. Provide respectively the
data, model, training and mefisto options to apply for the MEFISTO run. See \code{\link[MOFA2]{get_default_data_options}}, \code{\link[MOFA2]{get_default_model_options}},
\code{\link[MOFA2]{get_default_training_options}} and \code{\link[MOFA2]{get_default_mefisto_options}}.}

\item{only_common_samples}{Logical, whether only the samples present in all datasets should be returned.
Default value is \code{FALSE}.}
}
\value{
A \code{\link[MOFA2]{MOFA}} object.
}
\description{
Creates an object that can be used as input for the MEFISTO analysis implemented in the MOFA2 package. It contains the omics datasets
as well as features and samples metadata.
}
