% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{get_input_mixomics_supervised}
\alias{get_input_mixomics_supervised}
\title{Generate mixomics input data for supervised methods}
\usage{
get_input_mixomics_supervised(mo_data, group, datasets = names(mo_data))
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{group}{Character, the column name in the samples metadata data-frames to use as samples group
(use \code{\link{get_samples_metadata}} to view the samples information data-frame for each dataset).
This column should be either of type factor, character or integer.}

\item{datasets}{Character vector, the names of the datasets from \code{mo_data} to include in the analysis.}
}
\value{
A list, in which each element corresponds to one omics dataset, with samples as rows and features as columns.
The \code{Y} element is a factor vector with the outcome group of each sample.
}
\description{
Creates an object that can be used as input for the MixOmics package.
It contains the omics datasets restricted to common samples (with no missing group information) and
an outcome group for each sample.
}
\details{
Samples with missing values in the \code{group} column of the sample metadata will be removed from the dataset.
}
