% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofa.R
\name{get_input_mofa}
\alias{get_input_mofa}
\title{Generate MOFA input data}
\usage{
get_input_mofa(
  mo_data,
  datasets = names(mo_data),
  groups = NULL,
  options_list = NULL,
  only_common_samples = FALSE
)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{datasets}{Character vector, the names of the datasets from \code{mo_data} to include in the analysis.}

\item{groups}{Character, the column name in the samples metadata data-frames to use as groups
(use \code{\link{get_samples_metadata}} to view the samples metadata data-frame for each dataset).
WARNING: only use if you are familiar with MOFA and its use of groups. See \url{https://biofam.github.io/MOFA2/faq.html},
section "FAQ on the multi-group functionality".}

\item{options_list}{A named list. Should contain at most 3 elements, named 'data_options', 'model_options' and 'training_options'. Provide respectively the
data, model and training options to apply for the MOFA run. See \code{\link[MOFA2]{get_default_data_options}}, \code{\link[MOFA2]{get_default_model_options}}
and \code{\link[MOFA2]{get_default_training_options}}.}

\item{only_common_samples}{Logical, whether only the samples present in all datasets should be returned.
Default value is \code{FALSE}.}
}
\value{
A \code{\link[MOFA2]{MOFA}} object.
}
\description{
Creates an object that can be used as input for the MOFA analysis implemented in the MOFA2 package. It contains the omics datasets
as well as features and samples metadata.
}
