% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofa2.R
\name{get_input_mofa2}
\alias{get_input_mofa2}
\title{Generate input data for MOFA2 package}
\usage{
get_input_mofa2(
  mo_data,
  datasets,
  covariates,
  groups,
  options_list,
  only_common_samples
)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{datasets}{Character vector, the names of the datasets from \code{mo_data} to include in the analysis.}

\item{covariates}{Character or character vector of length 2, the column name(s) in the samples metadata
data-frames to use as continuous covariates. if NULL, creates an input object for MOFA. If not null,
creates an input object for MEFISTO.}

\item{groups}{Character, the column name in the samples metadata data-frames to use as group.}

\item{options_list}{A named list. Should contain at most 3 or 4 elements (depending on whether the input is for
MOFA or MEFISTO), named 'data_options', 'model_options', 'training_options' and 'mefisto_options' (latter only
for MEFISTO input).}

\item{only_common_samples}{Logical, whether only the samples present in all datasets should be returned.
Default value is \code{FALSE}.}
}
\value{
A \code{\link[MOFA2]{MOFA}} object.
}
\description{
Creates an object that can be used as input for the MOFA or MEFISTO analysis implemented in the MOFA2 package. It contains the omics datasets
as well as features and samples metadata. Should not be called directly; instead use \link{get_input_mofa} or \link{get_input_mefisto}.
}
