% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicspls.R
\name{get_input_omicspls}
\alias{get_input_omicspls}
\title{Generate omicsPLS input data}
\usage{
get_input_omicspls(mo_data, datasets = names(mo_data), scale_data = FALSE)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{datasets}{Character vector of length 2, the names of the datasets from \code{mo_data} to include in the analysis.}

\item{scale_data}{Boolean, should the datasets be scaled? Default value is \code{FALSE}.}
}
\value{
A list, in which each element corresponds to one omics dataset, with samples as rows and features as columns.
}
\description{
Creates an object that can be used as input for the omicsPLS package.
It contains the omics datasets restricted to common samples. Each dataset
is feature-centred.
}
