% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{get_latent_dimensions}
\alias{get_latent_dimensions}
\title{Get latent dimensions levels from dimension reduction output}
\usage{
get_latent_dimensions(method_output)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}
}
\value{
A character vector giving the labels of the latent dimensions.
}
\description{
Extracts the latent dimension levels from the output of a dimension reduction method.
}
