% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_running_time.R
\name{get_method_functions}
\alias{get_method_functions}
\title{Get functions used for each integration method}
\usage{
get_method_functions()
}
\value{
A named list of character vectors, each element corresponding to the
name of the functions that are used for data integration through a
particular method (e.g. DIABLO, MOFA, etc).
}
\description{
Get functions used for each integration method
}
