% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_get.R
\name{get_output}
\alias{get_output}
\alias{get_output_pca}
\alias{get_output_splsda}
\alias{get_output_spls}
\alias{get_output_diablo}
\alias{get_output_mofa2}
\alias{get_output_so2pls}
\title{Extract output of integration method in standard format}
\usage{
get_output(method_output, use_average_dimensions = TRUE)

get_output_pca(method_output)

get_output_splsda(method_output)

get_output_spls(method_output, use_average_dimensions = TRUE)

get_output_diablo(method_output, use_average_dimensions = TRUE)

get_output_mofa2(method_output)

get_output_so2pls(method_output, use_average_dimensions = TRUE)
}
\arguments{
\item{method_output}{The output of an integration method.}

\item{use_average_dimensions}{Logical, should the (weighted) average of the
samples scores for each latent dimension across the datasets be used? If
\code{FALSE}, a separate set of sample scores will be returned for each dataset
for each of the latent dimensions. Only applies to sPLS, DIABLO and sO2PLS
results. Default value is \code{TRUE}.}
}
\value{
An S3 object of class \code{output_dimension_reduction}, i.e. a named
list, with the following elements:
\itemize{
\item \code{features_weight}: tibble of features weight (loadings) for each latent
dimension, with columns \code{feature_id}, \code{dataset}, \code{latent_dimension},
\code{weight} (unscaled feature weight for the corresponding latent dimension),
\code{importance} (which corresponds to the scaled absolute weight, i.e. 1 =
feature has the maximum absolute weight for the corresponding latent
dimension and dataset, 0 = the feature was not selected for the
corresponding latent dimension)
\item \code{samples_score}: tibble of samples score for each latent component, with
columns \code{sample_id}, \code{latent_dimension}, \code{score} (unscaled samples score
for the corresponding latent dimension)
\item \code{variance_explained}: tibble of the fraction of variance explained by each
latent component for the relevant datasets.
}
}
\description{
Extract samples score and features weight from the result of an integration
method. The \code{get_output()} function provides a wrapper around the methods'
specific \verb{get_output_*()} functions.
}
