% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_pca.R
\name{get_pca_arguments}
\alias{get_pca_arguments}
\title{Extract arguments used in PCA run}
\usage{
get_pca_arguments(pca_result)
}
\arguments{
\item{pca_result}{The result of a PCA run on each of the datasets, computed
with the \code{\link[pcaMethods:pca]{pcaMethods::pca()}} function.}
}
\value{
A tibble with the following columns: \code{"Omics dataset"}, \code{"PCA method used"}, \code{"Number of Principal Components computed"}, \code{"Scaling applied"}
and \code{"Dataset centered"}.
}
\description{
Extracts the list of arguments used for each PCA run from a list of PCA
results, and formats them into a tibble.
}
