% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{get_samples}
\alias{get_samples}
\title{Get sample IDs from MultiDataSet}
\usage{
get_samples(mo_data)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
A named list, with one element per dataset, and each element is a
character vector of sample IDs.
}
\description{
Extract the list of sample IDs from each dataset in a MultiDataSet object.
}
