% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{get_samples_metadata_combined}
\alias{get_samples_metadata_combined}
\title{Get combined samples metadata data-frame from MultiDataSet}
\usage{
get_samples_metadata_combined(mo_data, only_common_cols = TRUE)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{only_common_cols}{Logical, whether to retain only common columns. If
\code{TRUE} (default value), only retain the columns that are present in the
samples metadata of all datasets. If \code{FALSE}, retain all columns from each
datasets' sample metadata.}
}
\value{
A data-frame of samples metadata.
}
\description{
Extracts the samples metadata data-frame (phenoData field) from each dataset
for a MultiDataSet object and combine them into one dataframe.
}
