% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_comparison.R
\name{get_samples_score_correlation}
\alias{get_samples_score_correlation}
\title{Get samples score correlation}
\usage{
get_samples_score_correlation(output_list)
}
\arguments{
\item{output_list}{List of integration methods output, each generated via the
\code{get_output()} function. If named, the names will be added at the beginning of
each latent dimension' label. If unnamed, the name of the integration method
will be used instead.}
}
\value{
A correlation matrix.
}
\description{
Constructs the correlation matrix between the samples score of the latent
dimensions obtained with different integration methods.
}
