% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{get_selected_features}
\alias{get_selected_features}
\title{Extract selected features}
\usage{
get_selected_features(
  method_output,
  latent_dimensions = NULL,
  datasets = NULL,
  mo_data = NULL
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{latent_dimensions}{Character vector of latent dimensions name. Default
value is \code{NULL} (top contributing features will be returned for all latent
dimensions).}

\item{datasets}{Character vector of datasets name. Default value is \code{NULL}
(top contributing features will be returned for all datasets).}

\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
A tibble containing one row per feature and latent dimension, giving
the weight and importance score of the feature for the corresponding latent dimension.
If \code{mo_data} is supplied, information about the features from
the features metadata will be added to the resulting table.
}
\description{
Extracts selected features from the output of an integration method. Only
features with a non-null weight for at least one latent dimension will be returned.
If a \code{MultiDataSet} object is supplied, information about the features from
the features metadata will be added to the resulting table.
}
