% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_running_time.R
\name{get_targets_running_time}
\alias{get_targets_running_time}
\alias{plot_running_time}
\title{Extract/plot running time of functions used for different integration methods}
\usage{
get_targets_running_time(
  target_patterns = c(sPLS = "^spls_", sO2PLS = "^so2pls_", MOFA = "^mofa_", DIABLO =
    "^diablo_"),
  patterns_to_methods = c("sPLS", "sO2PLS", "MOFA", "DIABLO"),
  target_exclude_patterns = NULL,
  method_functions = get_method_functions()
)

plot_running_time(
  target_patterns = c(sPLS = "^spls_", sO2PLS = "^so2pls_", MOFA = "^mofa_", DIABLO =
    "^diablo_"),
  patterns_to_methods = c("sPLS", "sO2PLS", "MOFA", "DIABLO"),
  target_exclude_patterns = NULL,
  method_functions = get_method_functions()
)
}
\arguments{
\item{target_patterns}{Named character vector, regex patterns used to extract
targets by their names.}

\item{patterns_to_methods}{Character vector of same length as
\code{target_patterns}, gives the integration method to which each pattern from
\code{target_patterns} corresponds.}

\item{target_exclude_patterns}{Character vector, regex used to exclude some
targets from the analysis.}

\item{method_functions}{Named list of character vector, giving for each
integration method (i.e. values in \code{patterns_to_methods}) the name of the
functions used in the pipeline to be considered when computing the running
time. Names should match values in \code{patterns_to_methods}.}
}
\value{
Either a tibble of running time for each function used as part of
each integration method (\code{get_targets_running_time}), or a ggplot
(\code{plot_running_time}).
}
\description{
In the plot, the coloured blocks in each bar each represent one of the steps
(targets) that correspond to an integration method. The text above each bar
shows the function that took the longest to run for each integration method.
}
\examples{
\dontrun{
## By default uses all targets whose name starts with "spls_" as part of the
## "sPLS" method, etc
get_targets_running_time()
plot_running_time()

## If we ran two versions of MOFA, e.g. one with supervised input and the
## other with unsupervised input
get_targets_running_time(target_exclude_patterns = "^mofa_unsupervised")
plot_running_time(target_exclude_patterns = "^mofa_unsupervised")

## If instead we want to compare the two MOFA run times
get_targets_running_time(
  target_patterns = c(
    "MOFA supervised" = "^mofa_(?!=unsupervised)",
    "MOFA unsupervised" = "^mofa_unsupervised"
  ),
  patterns_to_methods = c("MOFA", "MOFA")
)
plot_running_time(
  target_patterns = c(
    "MOFA supervised" = "^mofa_(?!=unsupervised)",
    "MOFA unsupervised" = "^mofa_unsupervised"
  ),
  patterns_to_methods = c("MOFA", "MOFA")
)
}
}
