% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{get_top_features}
\alias{get_top_features}
\title{Extract top features}
\usage{
get_top_features(
  method_output,
  n_features = 10,
  min_importance = NULL,
  latent_dimensions = NULL,
  datasets = NULL,
  mo_data = NULL
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{n_features}{Integer, the number of features to extract for each latent
dimension and dataset. Ignored if \code{min_importance} is set. Default value is \code{10}.
Will include all ties.}

\item{min_importance}{Numeric value between 0 and 1, minimum importance score
used to select features. Default value is \code{NULL}, i.e. the top \code{n_features}
features are selected instead.}

\item{latent_dimensions}{Character vector of latent dimensions name. Default
value is \code{NULL} (top contributing features will be returned for all latent
dimensions).}

\item{datasets}{Character vector of datasets name. Default value is \code{NULL}
(top contributing features will be returned for all datasets).}

\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
A tibble containing one row per feature and latent dimension, giving
the weight and importance score of the feature for the corresponding latent dimension.
If \code{mo_data} is supplied, information about the features from
the features metadata will be added to the resulting table.
}
\description{
Extracts the features with highest contribution to the latent dimensions
constructed by an integration method. Can retain a specific number of top
contributing features for each dataset and latent dimension, or all features
above a minimum importance score.
}
