% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{get_transformed_data}
\alias{get_transformed_data}
\title{Get MultiDataSet with transformed data}
\usage{
get_transformed_data(mo_data, transformation_result)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{transformation_result}{A list in which each element is the result of
a transformation applied to a different dataset, computed with the
\code{\link{transform_dataset}} function.}
}
\value{
A \code{\link[MultiDataSet]{MultiDataSet-class}} object, for which the
assay of each dataset is the imputed dataset.
}
\description{
Replace the original datasets with transformed datasets in a MultiDataSet
object from the results of transformations applied to the datasets.
}
