% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hclust_matrix_rows}
\alias{hclust_matrix_rows}
\title{Hierarchical clustering of matrix rows}
\usage{
hclust_matrix_rows(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
A dendrogram.
}
\description{
Performs a hierarchical clustering on the rows of a matrix. Code inspired
by ComplexHeatmap package.
}
