% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_dataset_csv}
\alias{import_dataset_csv}
\title{Import a dataset from a csv file}
\usage{
import_dataset_csv(file, col_id, features_as_rows = TRUE, ...)
}
\arguments{
\item{file}{Character, path to the dataset csv file.}

\item{col_id}{Character, the name of the column in the file that contains the
ID of the rows (i.e. feature IDs if \code{features_as_rows} is \code{TRUE}, or sample
IDs if \code{features_as_rows} is \code{FALSE}).}

\item{features_as_rows}{Logical, do the rows in the file correspond to
features? Default value is \code{TRUE}, i.e. the file contains features as rows
and samples as columns.}

\item{...}{Further arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\value{
A matrix with the samples as columns and the features as rows.
Feature IDs are used as row names and sample IDs as column names.
}
\description{
Reads a csv file and returns a matrix in which the rows corresponds to
features (e.g. markers, genes, phenotypes...) and the columns correspond to
samples/observations.
}
\examples{
\dontrun{
data_geno <- import_dataset_csv(
  "genotype_dataset.csv",
  col_id = "Marker",
  features_as_rows = TRUE
)
data_pheno <- import_dataset_csv(
  "phenotype_dataset.csv",
  col_id = "Sample",
  features_as_rows = FALSE
)
}
}
