% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_fmetadata_csv}
\alias{import_fmetadata_csv}
\title{Import feature metadata from a csv file}
\usage{
import_fmetadata_csv(file, col_id, ...)
}
\arguments{
\item{file}{Character, path to the dataset csv file.}

\item{col_id}{Character, the name of the column in the file that contains the
feature IDs.}

\item{...}{Further arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\value{
A data-frame with the features as rows and features information as
columns. Feature IDs are used as row names.
}
\description{
Reads a csv file and returns a dataframe in which the rows correspond to
features (e.g. markers, genes, phenotypes...) and columns correspond to
information about the features. Non-ASCII characters are replaced with ASCII
equivalents (using the stringi and textclean packages).
}
\examples{
\dontrun{
geno_info_features <- import_fmetadata_csv(
  "genotype_features_info.csv",
  col_id = "Marker"
)
}
}
