% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_fmetadata_csv_factory}
\alias{import_fmetadata_csv_factory}
\title{Target factory for csv features metadata import}
\usage{
import_fmetadata_csv_factory(files, col_ids, target_name_suffixes)
}
\arguments{
\item{files}{Character vector, vector of paths to the features metadata
csv files.}

\item{col_ids}{Character vector, the name of the column in each file that
contains the features ID.}

\item{target_name_suffixes}{Character vector, a suffix to add to the name of
the targets created by this target factory for each dataset.}
}
\value{
A list of target objects. For example, with two files to import and
\code{target_name_suffixes = c("geno", "transcripto")},
the factory returns the following targets:
\itemize{
\item \code{fmetadata_file_geno} and \code{fmetadata_file_transcripto}: targets tracking
the genomics and transcriptomics features metadata files, respectively.
\item \code{fmetadata_geno} and \code{fmetadata_transcripto}: targets that import the
genomics and transcriptomics features metadata dataset.
}
}
\description{
Creates a list of targets that track each file and import the features
metadata from each csv file.
}
\examples{
\dontrun{
## in the _targets.R
library(MOITestPipeline)

list(
  import_fmetadata_csv_factory(
    c(
      "data/genotype_fmetadata.csv",
      "data/rnaseq_fmetadata.csv"
    ),
    col_ids = c("Marker", "Info"),
    target_name_suffixes = c("geno", "transcripto")
  )
)
}
}
