% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_fmetadata_gff_factory}
\alias{import_fmetadata_gff_factory}
\title{Target factory for GFF/GTF features metadata import}
\usage{
import_fmetadata_gff_factory(
  files,
  feature_types,
  add_fieldss,
  target_name_suffixes
)
}
\arguments{
\item{files}{Character vector, vector of paths to the samples metadata GFF
or GTF files.}

\item{feature_types}{Character vector, the type of features to extract from
each annotation file. Currently supports \code{'genes'} and \code{'transcripts'}.}

\item{add_fieldss}{List, where each element is a character vector of  field
names in each GFF/GTF file to extract that are not imported by default.
If a character vector is provided, will be used for all files to read in.}

\item{target_name_suffixes}{Character vector, a suffix to add to the name of
the targets created by this target factory for each dataset.}
}
\value{
A list of target objects. For example, with two files to import and
\code{target_name_suffixes = c("geno", "transcripto")}, the factory returns the
following targets:
\itemize{
\item \code{fmetadata_file_geno} and \code{fmetadata_file_transcripto}: targets tracking
the genomics and transcriptomics annotation files, respectively.
\item \code{fmetadata_geno} and \code{fmetadata_transcripto}: targets that import the
genomics and transcriptomics features metadata datasets, respectively.
}
}
\description{
Creates a list of targets that track each file and import the features
metadata from each GFF/GTF file.
}
\examples{
\dontrun{
## in the _targets.R
library(MOITestPipeline)

list(
  import_fmetadata_gff_factory(
    c(
      "data/annotation.gff",
      "data/annotationv2.gtf"
    ),
    feature_types = c("genes", "transcripts"),
    add_fieldss = list(
      c("gene_name", "gene_custom_ID"),
      c("transcript_name")
    ),
    target_name_suffixes = c("geno", "transcripto")
  )
)
}
}
