% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_smetadata_csv}
\alias{import_smetadata_csv}
\title{Import samples metadata from a csv file}
\usage{
import_smetadata_csv(file, col_id, ...)
}
\arguments{
\item{file}{Character, path to the dataset csv file.}

\item{col_id}{Character, the name of the column in the file that contains
the ID of the rows (i.e. sample IDs).}

\item{...}{Further arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\value{
A data-frame with the samples as rows and the samples properties as
columns. Sample IDs are used as rownames.
}
\description{
Reads a csv file and returns a dataframe in which the rows
correspond to features (e.g. markers, genes, phenotypes...) and columns
correspond to information about the features.
}
\examples{
\dontrun{
samples_information <- import_smetadata_csv(
  "samples_information.csv",
  col_id = "Sample"
)
}
}
