% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_smetadata_csv_factory}
\alias{import_smetadata_csv_factory}
\title{Target factory for csv samples metadata import}
\usage{
import_smetadata_csv_factory(files, col_ids, target_name_suffixes)
}
\arguments{
\item{files}{Character vector, vector of paths to the samples metadata csv
files.}

\item{col_ids}{Character vector, the name of the column in each file that
contains the ID of the rows (i.e. the sample IDs).}

\item{target_name_suffixes}{Character vector, a suffix to add to the name of
the targets created by this target factory for each dataset.}
}
\value{
A list of target objects. For example, with two files to import and
\code{target_name_suffixes = c("geno", "transcripto")}, the factory returns the
following targets:
\itemize{
\item \code{smetadata_file_geno} and \code{smetadata_file_transcripto}: targets tracking
the genomics and transcriptomics samples metadata files, respectively.
\item \code{smetadata_geno} and \code{smetadata_transcripto}: targets that import the
genomics and transcriptomics samples metadata datasets, respectively.
}
}
\description{
Creates a list of targets that track each file and import the samples
metadata from each csv file.
}
\examples{
\dontrun{
## in the _targets.R
library(MOITestPipeline)

list(
  import_smetadata_csv_factory(
    c(
      "data/genotype_smetadata.csv",
      "data/rnaseq_smetadata.csv"
    ),
    col_ids = c("Sample", "SampleIDs"),
    target_name_suffixes = c("geno", "transcripto")
  )
)
}
}
