% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_equal_or_null}
\alias{is_equal_or_null}
\title{Check null or equality}
\usage{
is_equal_or_null(x, val)
}
\arguments{
\item{x}{The object to test.}

\item{val}{Value to compare to}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Tests whether an object is \code{NULL} or equal to some value.
}
