% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{join_features_metadata}
\alias{join_features_metadata}
\title{Join feature metadata to table}
\usage{
join_features_metadata(df, mo_data)
}
\arguments{
\item{df}{Data-frame of tibble with a column \code{feature_id} containing feature
IDs.}

\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
The \code{df} table with additional columns containing information about
the features from the features metadata table.
}
\description{
Adds features metadata information to a table containing feature IDs.
}
