% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{join_samples_metadata}
\alias{join_samples_metadata}
\title{Join samples metadata to table}
\usage{
join_samples_metadata(df, mo_data, datasets = NULL)
}
\arguments{
\item{df}{Data-frame or tibble with a column \code{id} containing sample IDs.}

\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{datasets}{Character vector, name(s) of datasets from which the samples
metadata should be extracted. If \code{NULL} (default value), information from
all datasets will be used.}
}
\value{
The \code{df} table with additional columns containing information about
the samples from the samples metadata table.
}
\description{
Adds samples metadata information to a table containing sample IDs.
}
