% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_evaluation.R
\name{make_feature_sets_from_df}
\alias{make_feature_sets_from_df}
\title{Makes list of feature sets from data-frame}
\usage{
make_feature_sets_from_df(annotation_df, col_id, col_set)
}
\arguments{
\item{annotation_df}{A data-frame of feature annotation in long format, with
at least a column of feature ID and a column giving the set to which the
feature belongs. If a feature belongs to more than one set, there should be a
row for each of these sets.}

\item{col_id}{Character, name of the column in \code{annotation_df} data-frame that
contains the features ID.}

\item{col_set}{Character, name of the column \code{annotation_df} data-frame that
contains the sets ID.}
}
\value{
a named list, where each element corresponds to a set, and contains
a vector of features ID of all features belonging to that set.
}
\description{
Creates a list of feature sets from an annotation data-frame.
}
