% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_classes_eSet.R
\docType{class}
\name{MetabolomeSet}
\alias{MetabolomeSet}
\alias{MetabolomeSet-class}
\title{Class to contain objects describing high-throughput metabolomics assays.}
\description{
Container for high-throughput metabolomics assays and experimental metadata.
MetabolomeSet class is derived from \code{\link[Biobase:class.eSet]{Biobase::eSet()}}.
}
