% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofa.R
\name{mofa_get_weights}
\alias{mofa_get_weights}
\title{Get feature weights from MOFA object}
\usage{
mofa_get_weights(
  object,
  views = "all",
  factors = "all",
  abs = FALSE,
  scale = "none",
  as.data.frame = TRUE
)
}
\arguments{
\item{object}{A trained MOFA object.}

\item{views}{Character or integer vector, name or index of the views (i.e. datasets) for which the feature weights
should be extracted. Default value is "all", i.e. all datasets are considered.}

\item{factors}{Character or integer vector, name or index of the factors for which the feature weights
should be extracted. Default value is "all", i.e. all factors are considered.}

\item{abs}{Logical, should the absolute value of the weights be returned? Default value is \code{FALSE}.}

\item{scale}{Character, the type of scaling that should be performed on the feature weights. Possible
values are \code{'none'}, \code{'by_view'}, \code{'by_factor'} or \code{'overall'} (see Details). Default value is \code{'none'}.}

\item{as.data.frame}{Logical, whether the function should return a long data-frame instead of a list
of matrices. Default value is \code{TRUE}.}
}
\value{
By default, returns a tibble with columns \code{view}, \code{feature}, \code{factor}, \code{value}. Alternatively, if
\code{as.data.frame = FALSE}, returns a list of matrices, one per view, with features as rows and factors as columns.
}
\description{
Extracts the feature weights from a trained MOFA or MEFISTO model (from the MOFA2 package).
}
\details{
Scaling options:
\itemize{
\item \code{scale = 'none'}: no scaling performed;
\item \code{scale = 'by_view'}: weights are divided by the maximum absolute weight in the corresponding view/dataset;
\item \code{scale = 'by_factor'}: weights are divided by the maximum absolute weight in the corresponding factor;
\item \code{scale = 'overall'}: weights are divided by the maximum absolute weight across all views/datasets and
factors considered.
}
}
