% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofa.R
\name{mofa_plot_cor_covariates}
\alias{mofa_plot_cor_covariates}
\title{Plots the correlation between factors and covariates for MOFA}
\usage{
mofa_plot_cor_covariates(
  mofa_output,
  covariates = NULL,
  show_cor = TRUE,
  min_show_cor = 0.2,
  round_cor = 2,
  factor_as_col = TRUE
)
}
\arguments{
\item{mofa_output}{The output from \code{\link[MOFA2]{run_mofa}}.}

\item{covariates}{Character vector, the covariates to use for the plot. If \code{NULL},
all covariates retrieved via \code{colnames(MOFA2::samples_metadata(mofa_output))}
(except \code{group}, \code{id} and \code{sample}) will be used. Default value is \code{NULL}.}

\item{show_cor}{Logical, should the correlation values be added to the plot?
Default value is \code{TRUE}.}

\item{min_show_cor}{Numeric, the minimum value below which correlation
coefficients values are not added to the plot (i.e. only a circle will
appear for these values but no text). Ignored if \code{show_cor} is \code{FALSE}.
Default value is 0.2.}

\item{round_cor}{Integer, how many decimal places to show for the correlation
coefficients. Ignored if \code{show_cor} is \code{FALSE}. Default value is 2.}

\item{factor_as_col}{Logical, should the factors be represented in columns?
If \code{FALSE}, will be represented in rows instead. Default value is \code{TRUE}.}
}
\value{
A ggplot.
}
\description{
Plots the Pearson correlation between MOFA latent factors and covariates
(obtained from the samples metadata). This function provides the \code{ggplot2}
version of the plot created with \code{\link[MOFA2]{correlate_factors_with_covariates}}
(with the \code{plot} parameter set to \code{'r'}).
}
