% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{options_list_as_tibble}
\alias{options_list_as_tibble}
\title{Returns options list as a tibble}
\usage{
options_list_as_tibble(options_list)
}
\arguments{
\item{options_list}{A named list, where each element corresponds to one
option or parameter and the name of the element corresponds to the name of
the option/parameter.}
}
\value{
a tibble, with the \code{Parameter} column giving the list of the options
or parameters, and the \code{Value} column giving the values of the
corresponding option or parameter.
}
\description{
Transforms a list of options (or parameters) into a tibble with the name of
the options (parameters) in one column, and their value in a second column.
Vector values are collapsed to span only one column.
}
