% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{perf_splsda}
\alias{perf_splsda}
\title{Assess optimal number of components for sPLS-DA on omics dataset from MultiDataSet object}
\usage{
perf_splsda(
  splsda_input,
  ncomp_max = 5,
  validation = "Mfold",
  folds = 5,
  nrepeat = 50,
  measure = "BER",
  distance = "centroids.dist",
  cpus = 1,
  progressBar = TRUE
)
}
\arguments{
\item{splsda_input}{Input for the sPLS-DA functions from mixOmics, created with \code{\link[=get_input_splsda]{get_input_splsda()}}.}

\item{ncomp_max}{Integer, the maximum number of latent components to test when estimating the number of
latent components to use. Default value is \code{5}.}

\item{validation}{Character, which cross-validation method to use, can be one of \code{"Mfold"} or \code{"loo"}
(see \code{\link[mixOmics:perf]{mixOmics::perf()}}). Default value is \code{"Mfold"}.}

\item{folds}{Integer, number of folds to use in the M-fold cross-validation (see \code{\link[mixOmics:perf]{mixOmics::perf()}}).
Default value is 5.}

\item{nrepeat}{Integer, number of times the cross-validation is repeated (see \code{\link[mixOmics:perf]{mixOmics::perf()}}).}

\item{measure}{Performance measure used to select the optimal value of \code{ncomp}, can be one of \code{"BER"} or \code{"overall"}
(see \code{\link[mixOmics:perf]{mixOmics::perf()}}).
Default value is \code{"BER"}.}

\item{distance}{Distance metric used to select the optimal value of \code{ncomp}, can be one of \code{"max.dist"},
\code{"centroids.dist"} or \code{"mahalanobis.dist"} (see \code{\link[mixOmics:perf]{mixOmics::perf()}}). Default value is \code{"centroids.dist"}.}

\item{cpus}{Integer, number of cpus to use.}

\item{progressBar}{Logical, whether to display a progress bar during the optimisation of \code{ncomp}. Default
value is \code{TRUE}.}
}
\value{
A list as per the output of the \code{\link[mixOmics:perf]{mixOmics::perf()}} function, with the following additional elements:
\itemize{
\item \code{dataset_name}: the name of the dataset analysed;
\item \code{group}: column name in the samples information data-frame used as samples group;
\item \code{optim_ncomp}: the optimal number of latent components as per the \code{measure} and \code{distance} specified;
\item \code{optim_measure}: the measure used to select the optimal number of latent components;
\item \code{optim_distance}: the distance metric used to select the optimal number of latent components.
}
In addition, the name of the dataset analysed and the column name in the samples information data-frame
used as samples group as stored as attributes \code{dataset_name} and \code{group}, respectively.
}
\description{
Performs cross-validation for a PLS-DA run (implemented in the \code{mixOmics} package) on an omics dataset from a
\code{MultiDataSet} object. This allows to estimate the optimal number of latent components to construct.
This is intended for feature preselection in the omics dataset (see examples below).
}
\details{
This function uses the \code{\link[mixOmics]{plsda}} and  \code{\link[mixOmics]{perf}}
function from the \code{mixOmics} package.
}
