% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_classes_eSet.R
\docType{class}
\name{PhenotypeSet}
\alias{PhenotypeSet}
\alias{PhenotypeSet-class}
\title{Class to contain objects describing phenotypic assays.}
\description{
Container for phenotypic assays and experimental metadata. PhenotypeSet class
is derived from \code{\link[Biobase:class.eSet]{Biobase::eSet()}}.
}
