% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_correlation_matrix_full}
\alias{plot_correlation_matrix_full}
\title{Plots a full correlation matrix (corrplot-style)}
\usage{
plot_correlation_matrix_full(
  mat,
  rows_title = NULL,
  cols_title = NULL,
  title = NULL,
  show_cor = TRUE,
  min_show_cor = 0.2,
  round_cor = 2
)
}
\arguments{
\item{mat}{Correlation matrix to plot.}

\item{rows_title}{Character, title for rows. Default value is \code{NULL}.}

\item{cols_title}{Character, title for cols. Default value is \code{NULL}.}

\item{title}{Character, title of the plot. Default value is \code{NULL}.}

\item{show_cor}{Logical, should the correlation values be added to the plot?
Default value is \code{TRUE}.}

\item{min_show_cor}{Numeric, the minimum value below which correlation
coefficients values are not added to the plot (i.e. only a circle will
appear for these values but no text). Ignored if \code{show_cor} is \code{FALSE}.
Default value is 0.2.}

\item{round_cor}{Integer, how many decimal places to show for the correlation
coefficients. Ignored if \code{show_cor} is \code{FALSE}. Default value is 2.}
}
\value{
a ggplot.
}
\description{
Generates a plot of a correlation matrix in the style of the corrplot
package.
}
