% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_plots.R
\name{plot_data_heatmap}
\alias{plot_data_heatmap}
\title{Plots omics data as heatmap}
\usage{
plot_data_heatmap(
  mo_data,
  features,
  center = FALSE,
  scale = FALSE,
  samples = NULL,
  only_common_samples = FALSE,
  samples_info = NULL,
  features_info = NULL,
  colours_list = NULL,
  label_cols = NULL,
  truncate = NULL,
  legend_title_size = 10,
  legend_text_size = 10,
  ...
)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{features}{Character vector, the ID of features to show in the plot.}

\item{center}{Logical, whether the data should be centered (feature-wise).
Default value is \code{FALSE}.}

\item{scale}{Logical, whether the data should be scaled (feature-wise).
Default value is \code{FALSE}.}

\item{samples}{Character vector, the ID of samples to include in the plot. If
\code{NULL} (default), all samples will be used.}

\item{only_common_samples}{Logical, whether only samples that are present in
all datasets should be plotted. Default value is \code{FALSE}.}

\item{samples_info}{Character vector, column names from the samples metadata
tables of the datasets to be represented in the plot as samples annotation.}

\item{features_info}{Named list of character vectors, where each element
corresponds to a dataset, and gives the column names from the features
metadata of the dataset to be represented in the plot as features
annotation. The names of the list must correspond to dataset names in the
\code{mo_data} object.}

\item{colours_list}{Named list, where each element gives the colour palette
to use in the samples or features annotation. Names must match values in
\code{samples_info} vector and elements of \code{features_info} list. For continuous
palettes, must use \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}} function (see the
\href{https://jokergoo.github.io/ComplexHeatmap-reference/book/heatmap-annotations.html#simple-annotation}{ComplexHeatmap reference book}).}

\item{label_cols}{Character or named list of character, giving for each
dataset the name of the column in the corresponding features metadata to
use as label. If one value, will be used for all datasets. If list, the
names must correspond to the names of the datasets in \code{mo_data}. If a
dataset is missing from the list or no value is provided, feature IDs will
be used as labels. Alternatively, use \code{feature_id} to get the feature IDs
as labels.}

\item{truncate}{Integer, width to which the labels should be truncated (to
avoid issues with very long labels in plots). If \code{NULL} (default value), no
truncation will be performed.}

\item{legend_title_size}{Integer, size in points of legend title.}

\item{legend_text_size}{Integer, size in points of legend elements text.}

\item{...}{Additional arguments passed to the \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}
function.}
}
\value{
A \link[ComplexHeatmap:Heatmap-class]{ComplexHeatmap::Heatmap} object.
}
\description{
For a given set of features, plots their value across the samples as a
heatmap.
}
\examples{
\dontrun{
## Selecting at random 3 features from each dataset
random_features <- get_features(mo_set) |>
  map(sample, size = 3, replace = FALSE) |>
  unlist() |>
  unname()

plot_data_heatmap(
  mo_set,
  random_features,
  center = TRUE,
  scale = TRUE,
  show_column_names = FALSE,
  only_common_samples = TRUE,
  samples_info = c("status", "day_on_feed"),
  features_info = c("chromosome"),
  colours_list = list(
    "status" = c("Control" = "gold", "BRD" = "navyblue"),
    "day_on_feed" = colorRamp2(c(5, 65), c("white", "pink3"))
  ),
  label_cols = list(
     "rnaseq" = "Name",
    "metabolome" = "name"
  )
 )
}
}
