% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_plots.R
\name{plot_density_data}
\alias{plot_density_data}
\title{Per-dataset density plot for MultiDataSet object}
\usage{
plot_density_data(
  mo_data,
  datasets = names(mo_data),
  combined = TRUE,
  scales = "fixed"
)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{datasets}{Character vector, names of the datasets to include in the
plot. By default, all datasets are included.}

\item{combined}{Logical, should the different datasets be represented in the
same plot? If \code{FALSE} (default value), each dataset will be represented in
its own subplot. Default value is \code{TRUE}.}

\item{scales}{Character, how should the axes be plotted if \code{combined = FALSE}. Can be either \code{'fixed'}, i.e. the same limits will be applied to
the axes of each subplot; or \code{'free'}, i.e. the axis limits will be adapted
to each subplot. Ignored if \code{combined = TRUE}. Default value is \verb{'fixed}.}
}
\value{
A ggplot.
}
\description{
Displays a density plot of the values in each dataset of a MultiDataSet
object.
}
