% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{plot_feature_preselection_cov}
\alias{plot_feature_preselection_cov}
\title{Diagnostics plots for COV-based feature preselection}
\usage{
plot_feature_preselection_cov(cov_list)
}
\arguments{
\item{cov_list}{A list with the result from the COV calculation for each dataset
to be filtered, as returned by the \code{\link{select_features_cov}} function.}
}
\value{
A ggplot.
}
\description{
Displays the COV distribution across all features in the original (i.e. non-filtered) datasets,
with a vertical red line showing the cut-off used by the preselection function.
}
