% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{plot_feature_preselection_mad}
\alias{plot_feature_preselection_mad}
\title{Diagnostics plots for MAD-based feature preselection}
\usage{
plot_feature_preselection_mad(mad_list)
}
\arguments{
\item{mad_list}{A list with the result from the MAD calculation for each dataset
to be filtered, as returned by the \code{\link{select_features_mad}} function.}
}
\value{
A ggplot.
}
\description{
Displays the MAD distribution across all features in the original (i.e. non-filtered) datasets,
with a vertical red line showing the cut-off used by the preselection function.
}
