% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{plot_feature_preselection_splsda}
\alias{plot_feature_preselection_splsda}
\title{Diagnostics plots for sPLS-DA-based feature preselection}
\usage{
plot_feature_preselection_splsda(
  perf_splsda_res,
  measure = NULL,
  distance = NULL
)
}
\arguments{
\item{perf_splsda_res}{A list with the result from the \code{\link{perf_splsda}} for each dataset
to be filtered.}

\item{measure}{Which measure(s) should be displayed? Can be one of \code{"BER"}
or \code{"overall"}. If NULL, all measures will be displayed. Default value is \code{NULL}.}

\item{distance}{Which measure(s) should be displayed? Can be one of \code{"max.dist"},
\code{"centroids.dist"} or \code{"mahalanobis.dist"}. If NULL, all measures will be displayed. Default value is \code{NULL}.}
}
\value{
A ggplot.
}
\description{
Displays the PLS-DA classification performance across different number of
latent components for each prefiltered dataset. The classification error rates are computed with different measures
(column facets) and different distance metrics (colours). A vertical grey bar represents for each dataset the number
of latent components selected for the feature preselection step. In addition, a circle highlights the measure and
distance metric used to select the number of latent component.
}
