% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plot.R
\name{plot_features_weight_distr}
\alias{plot_features_weight_distr}
\title{Plots features weight distribution}
\usage{
plot_features_weight_distr(
  method_output,
  latent_dimensions = NULL,
  datasets = NULL,
  features_metric = c("signed_importance", "weight", "importance"),
  top_n = 0,
  mo_data = NULL,
  label_cols = NULL,
  truncate = NULL,
  text_size = 2.5
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{latent_dimensions}{Character vector giving the latent dimensions to display.
Default value is \code{NULL}, i.e. all latent dimensions will be shown.}

\item{datasets}{Character vector giving the datasets to display.
Default value is \code{NULL}, i.e. all datasets will be shown.}

\item{features_metric}{Character, which attribute should be plotted: can be
\code{'signed_importance'} (i.e. importance value but with the weight sign),
\code{'importance'} or \code{'weight'}. Default value is \code{'signed_importance'}.}

\item{top_n}{Integer, number of top features (in terms of importance) for which
the label should be shown. Default value is \code{0}.}

\item{mo_data}{A \code{MultiDataSet} object. Only used if \code{label_cols} is not
\code{NULL}.}

\item{label_cols}{Character or named list of character, giving for each
dataset the name of the column in the corresponding features metadata to
use as label. If one value, will be used for all datasets. If list, the
names must correspond to the names of the datasets in \code{mo_data}. If a
dataset is missing from the list or no value is provided, feature IDs will
be used as labels. Alternatively, use \code{feature_id} to get the feature IDs
as labels.}

\item{truncate}{Integer, width to which the labels should be truncated (to
avoid issues with very long labels in plots). If \code{NULL} (default value), no
truncation will be performed.}

\item{text_size}{Numeric, size of the feature labels.}
}
\value{
A \code{\link[patchwork]{patchwork}} of plots.
}
\description{
Plots the features weight or importance per dataset and latent dimension.
}
