% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_plots.R
\name{plot_meansd_data}
\alias{plot_meansd_data}
\title{Per-dataset mean-sd trend plots for MultiDataSet object}
\usage{
plot_meansd_data(
  mo_data,
  datasets = names(mo_data),
  by_rank = FALSE,
  colour_log10 = TRUE
)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}

\item{datasets}{Character vector, names of the datasets to include in the
plot. By default, all datasets are included.}

\item{by_rank}{Logical, should the x-axis display the rank of the features
(ordered by mean) rather than the features mean? Default value is \code{FALSE},
i.e. the x axis represents the mean of the features.}

\item{colour_log10}{Should the colour legend be on the log10 scale? Default
value is \code{TRUE}.}
}
\value{
A ggplot.
}
\description{
Displays for each dataset in a MultiDataSet object the trend between features
mean and standard deviation across all samples.
}
