% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_pca.R
\name{plot_samples_coordinates_pca}
\alias{plot_samples_coordinates_pca}
\title{Samples score plots for single-omics PCA}
\usage{
plot_samples_coordinates_pca(pca_result, pcs = NULL, datasets = NULL, ...)
}
\arguments{
\item{pca_result}{List of PCA results on each of the datasets, each computed
with the \code{\link[=run_pca]{run_pca()}} function.}

\item{pcs}{Integer vector or named list of integer vectors, the principal
components to display for each dataset. If integer vector (e.g. \code{1:5}),
will be used for all datasets. Alternatively, a different set of PCs can be
specified through a named list (e.g. \code{list('snps' = 1:4, 'rnaseq' = 1:5)}).
The length of the list must match the number of datasets to be displayed,
and the names must match the dataset names. Default value is \code{NULL}, i.e.
all principal components will be plotted for each dataset.}

\item{datasets}{Optional, character vector of datasets for which the plots
should be created.}

\item{...}{Other arguments passed to \code{\link[=plot_samples_score]{plot_samples_score()}}.}
}
\value{
A list of ggmatrix plots (or a single ggmatrix plot if \code{pcs} has
length 1).
}
\description{
Produces a pairwise samples score plot for the PCA run on each omics dataset,
using \code{\link[GGally:ggpairs]{GGally::ggpairs()}}.
}
\examples{
\dontrun{
## Default: plotting all PCs for all datasets
plot_samples_coordinates_pca(pca_result)

## Plotting only the first 3 PCs for each dataset
plot_samples_coordinates_pca(
  pca_result,
  pcs = 1:3
)

## Plotting the first 3 PCs for the genomics dataset, 4 PCs for the
## transcriptomics dataset, 5 PCs for the metabolomics dataset
plot_samples_coordinates_pca(
  pca_result,
  pcs = list(
    "snps" = 1:3,
    "rnaseq" = 1:4,
    "metabolome" = 1:5
  )
)

## Plotting the first 3 PCs for the genomics and transcriptomics datasets
plot_samples_coordinates_pca(
  pca_result,
  pcs = 1:3,
  datasets = c("snps", "rnaseq")
)

# Plotting the first 3 PCs for the genomics dataset and 4 PCs for the
## transcriptomics dataset (no plot for the metabolomics dataset)
plot_samples_coordinates_pca(
  pca_result,
  pcs = list(
    "snps" = 1:3,
    "rnaseq" = 1:4
  ),
  datasets = c("snps", "rnaseq")
)
}
}
