% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_plots.R
\name{plot_samples_upset}
\alias{plot_samples_upset}
\title{Upset plot of samples}
\usage{
plot_samples_upset(mo_data)
}
\arguments{
\item{mo_data}{A \link[MultiDataSet:MultiDataSet-class]{MultiDataSet::MultiDataSet} object.}
}
\value{
An upset plot.
}
\description{
Generates an upset plot to compare the samples present in each omics dataset.
}
