% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_pca.R
\name{plot_screeplot_pca}
\alias{plot_screeplot_pca}
\title{Screeplots for single-omics PCA}
\usage{
plot_screeplot_pca(pca_result, cumulative = FALSE, datasets = NULL)
}
\arguments{
\item{pca_result}{List of PCA runs result on each of the datasets, each
computed with the \code{\link[=run_pca]{run_pca()}} function.}

\item{cumulative}{Logical, should the cumulative variance be plotted? Default
is \code{FALSE}.}

\item{datasets}{Optional, character vector with the names of the datasets to
plot. If \code{NULL} (default value), all datasets will be plotted.}
}
\value{
A ggplot2 plot.
}
\description{
Produces a scree plot (percentage of variance explained by each principal
component) for the PCA run on each omics dataset, using ggplot2.
}
