% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plot.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Plot of variance explained}
\usage{
plot_variance_explained(
  method_output,
  datasets = NULL,
  latent_dimensions = NULL,
  ncol = NULL,
  free_y_axis = FALSE,
  cumulative = FALSE
)
}
\arguments{
\item{method_output}{Integration method output generated via the
\code{get_output()} function.}

\item{datasets}{Character vector giving the datasets to display. Default
value is \code{NULL}, i.e. all datasets will be shown.}

\item{latent_dimensions}{Character vector giving the latent dimensions to
display. Default value is \code{NULL}, i.e. all latent dimensions will be shown.}

\item{ncol}{Integer, number of columns in the faceted plot. Default value is
\code{NULL}.}

\item{free_y_axis}{Logical, whether the y-axis (representing the percentage
of variance) should have the same range for all datasets. Default value is
\code{FALSE}.}

\item{cumulative}{Logical, whether the cumulative percentage of variance
explained should be plotted. Default is \code{FALSE}.}
}
\value{
A ggplot.
}
\description{
Displays the percentage of variance explained by each latent dimension in the
output of a dimension reduction method for each dataset analysed.
}
