% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_x_continuous}
\alias{plot_x_continuous}
\title{Scatter plot function}
\usage{
plot_x_continuous(toplot, x, y, colour, shape, point_alpha = 1, add_se = TRUE)
}
\arguments{
\item{toplot}{Tibble with data to plot.}

\item{x}{Character, name of the column in \code{toplot} to use as x-axis.}

\item{y}{Character, name of the column in \code{toplot} to use as y-axis.}

\item{colour}{Character, name of the column in \code{toplot} to use as colour.}

\item{shape}{Character, name of the column in \code{toplot} to use as shape.}

\item{point_alpha}{Numeric between 0 and 1, the opacity of the points in the
plot (with 1 = fully opaque, and 0 = fully transparent). Default value is
\code{1}.}

\item{add_se}{Logical, should a confidence interval be drawn around the
smoothing curve for scatterplots? Default value is \code{TRUE}.}
}
\value{
A ggplot.
}
\description{
Creates a scatter plot
}
\details{
The function adds a loess curve to summarise the trend between the covariate
and the samples score. If \code{colour_by} is used, and the corresponding variable
is numeric, the loess curve will not take into account this variable. If
instead the \code{colour_by} variable is a character or factor, a loess curve will
be fitted separately for each category.
}
