% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_x_discrete}
\alias{plot_x_discrete}
\title{Violin plot function function}
\usage{
plot_x_discrete(
  toplot,
  x,
  y,
  colour,
  shape,
  point_alpha = 1,
  add_boxplot = TRUE
)
}
\arguments{
\item{toplot}{Tibble with data to plot.}

\item{x}{Character, name of the column in \code{toplot} to use as x-axis.}

\item{y}{Character, name of the column in \code{toplot} to use as y-axis.}

\item{colour}{Character, name of the column in \code{toplot} to use as colour.}

\item{shape}{Character, name of the column in \code{toplot} to use as shape.}

\item{point_alpha}{Numeric between 0 and 1, the opacity of the points in the
plot (with 1 = fully opaque, and 0 = fully transparent). Default value is
\code{1}.}

\item{add_boxplot}{Logical, should a boxplot be drawn on top of the points
for violin plots? Default value is \code{TRUE}.}
}
\value{
A ggplot.
}
\description{
Creates a violin plot
}
\details{
If \code{colour_by} is used, and the corresponding variable is numeric, the
violins and boxplots will not take into account this variable. If instead the
\code{colour_by} variable is a character or factor, a separate violin and boxplot
will be drawn for each category.
}
