% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_x_wrapper}
\alias{plot_x_wrapper}
\title{Wrapper to create plot}
\usage{
plot_x_wrapper(
  toplot,
  x,
  y,
  colour,
  shape,
  point_alpha = 1,
  add_se = TRUE,
  add_boxplot = TRUE,
  facet_wrap = NULL,
  ncol_wrap = NULL,
  facet_grid = NULL,
  scales_facet = "free_y"
)
}
\arguments{
\item{toplot}{Tibble with data to plot.}

\item{x}{Character, name of the column in \code{toplot} to use as x-axis.}

\item{y}{Character, name of the column in \code{toplot} to use as y-axis.}

\item{colour}{Character, name of the column in \code{toplot} to use as colour.}

\item{shape}{Character, name of the column in \code{toplot} to use as shape.}

\item{point_alpha}{Numeric between 0 and 1, the opacity of the points in the
plot (with 1 = fully opaque, and 0 = fully transparent). Default value is
\code{1}.}

\item{add_se}{Logical, should a confidence interval be drawn around the
smoothing curve for scatterplots? Default value is \code{TRUE}.}

\item{add_boxplot}{Logical, should a boxplot be drawn on top of the points
for violin plots? Default value is \code{TRUE}.}

\item{facet_wrap}{Character, name of the column in \code{toplot} to use for
faceting (using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}). Default is \code{NULL}.}

\item{ncol_wrap}{Integer, number of columns in the faceted plot if using
\code{facet_wrap}. Default value is \code{NULL}.}

\item{facet_grid}{Character vector of length 2, name of the columns in
\code{toplot} to use for row (first element) and column (second element)
faceting (using \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}). Will be ignored if \code{facet_wrap}
is not \code{NULL}. Default is \code{NULL}.}

\item{scales_facet}{Character, value to use for the \code{scales} argument of
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} or \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}. Default value is
\code{'free_y'}.}
}
\value{
A ggplot.
}
\description{
Wrapper around \code{\link[=plot_x_continuous]{plot_x_continuous()}} and \code{\link[=plot_x_discrete]{plot_x_discrete()}}, will choose
which one to use depending on whether the x-axis variable is continuous or
discrete.
}
