% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_evaluation.R
\name{reduce_feature_sets_data}
\alias{reduce_feature_sets_data}
\title{Reduce feature sets to match multi-omics dataset}
\usage{
reduce_feature_sets_data(feature_sets, mo_data, datasets = names(mo_data))
}
\arguments{
\item{feature_sets}{Named list, where each element corresponds to a feature set,
and contains a vector of features ID of all features belonging to that set.}

\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{datasets}{Character vector, the names of the datasets for which features
assignment should be checked. By default, all datasets in \code{mo_data} are considered.}
}
\value{
A feature sets list, i.e. named list where each element corresponds to
a feature set, containing the ID of all features that belong to this set and
that are present in the multi-omics dataset.
}
\description{
Removes from a list of feature sets all features that are not present in the
multi-omics dataset.
}
