% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diablo.R
\name{run_pairwise_pls}
\alias{run_pairwise_pls}
\title{Pairwise PLS datasets comparison}
\usage{
run_pairwise_pls(mixomics_data, datasets_name, ..., verbose = TRUE)
}
\arguments{
\item{mixomics_data}{A \code{mixOmics} input object created with \code{\link{get_input_mixomics_supervised}}.}

\item{datasets_name}{Character vector of length 2, the names of the two omics datasets to analyse.}

\item{...}{Additional parameters to be passed to the \code{\link[mixOmics]{pls}} function.}

\item{verbose}{Logical, should details be printed during execution? Default value is \code{TRUE}.}
}
\value{
A named list; each element is an object of class \code{\link[mixOmics]{pls}},
which provides the result of the PLS run. The name of the datasets analysed is stored as a character vector
in the \code{datasets_name} attribute.
}
\description{
Runs a Projection to Latent Structure (PLS) analysis on a pair of omics datasets, as per the mixOmics package.
}
\details{
Note only one latent component is computed as only the first latent component will be used to assess the
correlation between datasets.
}
\examples{
\dontrun{
run_pairwise_pls(mo_set, c("rnaseq", "metabolome"))
}
}
