% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_pca.R
\name{run_pca_matrix}
\alias{run_pca_matrix}
\title{Run PCA on matrix}
\usage{
run_pca_matrix(mat, n_pcs = 10, scale = "none", center = TRUE, method = NULL)
}
\arguments{
\item{mat}{Matrix of omics measurement, with features as rows and samples as
columns.}

\item{n_pcs}{numeric, number of Principal Components to compute. Default
value is 10.}

\item{scale}{character, type of scaling that should be applied to the dataset
before running the PCA. Should be one of \code{'none'}, \code{'pareto'}, \code{'vector'},
\code{'uv'} (see \code{\link[pcaMethods:pca]{pcaMethods::pca()}}). Default value is \code{'none'}.}

\item{center}{boolean, should the dataset be centred prior to running the
PCA? Default value is \code{TRUE}.}

\item{method}{character, type of PCA that should be applied to the dataset.
See \code{\link[pcaMethods:listPcaMethods]{pcaMethods::listPcaMethods()}}. for a list of available methods.
Default value is \code{'svd'} for datasets with no missing value, and \code{'nipals'}
for datasets with missing values.}
}
\value{
A \link[pcaMethods:pcaRes]{pcaMethods::pcaRes} object containing the result from the
PCA analysis. The attribute \code{dataset_name} specifies the name of the
dataset analysed.
}
\description{
Runs a Principal Component Analysis on an omics matrix, using the
\code{\link[pcaMethods:pca]{pcaMethods::pca()}} function.
}
