% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{select_features_cov}
\alias{select_features_cov}
\title{Select features based on Coefficient of Variation from MultiDataSet}
\usage{
select_features_cov(
  mo_data,
  dataset_name,
  to_keep_n = NULL,
  to_keep_prop = NULL,
  with_ties = TRUE
)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}

\item{dataset_name}{Character, name of the omics dataset on which to apply feature pre-selection.}

\item{to_keep_n}{Integer, the number of features to retain in the dataset. Should be less than the number of
features in the dataset. If \code{NULL} or \code{NA}, \code{to_keep_prop} will be used instead.}

\item{to_keep_prop}{Numeric, the proportion of features to retain in the dataset. Will be ignored if \code{to_keep_n}
is supplied. Value should be > 0 and < 1.}

\item{with_ties}{Should ties be kept together? If \code{TRUE}, may return more features than requested. Default
value is \code{TRUE}.}
}
\value{
A tibble with columns \code{feature_id}, \code{cov} and \code{selected} (logical, indicates whether the feature is selected
based on its COV value). In addition, the name of the dataset filtered is stored in the return object attribute
\code{dataset_name} (which can be accessed via \code{attr(res, "dataset_name")}).
}
\description{
Computes the Coefficient of Variation (COV) for each feature in an omics dataset
from a MultiDataSet object, and select features with the highest COV values. This
is a wrapper function around the \code{\link[=get_dataset_matrix]{get_dataset_matrix()}} and \code{\link[=select_features_cov_matrix]{select_features_cov_matrix()}} functions.
}
