% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{select_features_mad_matrix}
\alias{select_features_mad_matrix}
\title{Select features based on Median Absolute Deviation from matrix}
\usage{
select_features_mad_matrix(
  mat,
  to_keep_n = NULL,
  to_keep_prop = NULL,
  with_ties = TRUE
)
}
\arguments{
\item{mat}{Matrix of omics measurement, with features as rows and samples as columns.}

\item{to_keep_n}{Integer, the number of features to retain in the dataset. Should be less than the number of
features in the dataset. If \code{NULL} or \code{NA}, \code{to_keep_prop} will be used instead.}

\item{to_keep_prop}{Numeric, the proportion of features to retain in the dataset. Will be ignored if \code{to_keep_n}
is supplied. Value should be > 0 and < 1.}

\item{with_ties}{Should ties be kept together? If \code{TRUE}, may return more features than requested. Default
value is \code{TRUE}.}
}
\value{
A tibble with columns \code{feature_id}, \code{mad} and \code{selected} (logical, indicates whether the feature is selected
based on its MAD value). In addition, the name of the dataset filtered is stored in the return object attribute
\code{dataset_name} (which can be accessed via \code{attr(res, "dataset_name")}).
}
\description{
Computes the Median Absolute Deviation (MAD) for each feature in an omics dataset
from a MultiDataSet object, and select features with the highest MAD values.
}
