% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_comparison.R
\name{show_consensus_metrics}
\alias{show_consensus_metrics}
\title{Illustrates importance consensus metrics}
\usage{
show_consensus_metrics(
  metrics = c("min", "harmonic", "geometric", "product", "average", "l2", "max")
)
}
\arguments{
\item{metrics}{Character vector of metrics to show. Should be valid values for
the \code{metric} argument of \code{\link[=consensus_importance_metric]{consensus_importance_metric()}}, i.e. "min", "max",
"average", "product", "l2", "geometric", "harmonic".}
}
\value{
A ggplot.
}
\description{
Plots a heatmap to illustrate the behaviour of different importance consensus metrics.
}
