% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_compare_samples_joint_components}
\alias{so2pls_compare_samples_joint_components}
\title{Compares sO2PLS samples joint component scores between the two datasets}
\usage{
so2pls_compare_samples_joint_components(so2pls_res, components = NULL, ...)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}

\item{components}{Optional, an integer vector with the joint components
that should be plotted. Default is \code{NULL}, i.e. all joint components are
represented.}

\item{...}{Further arguments passed to \code{\link{plot_samples_score_pair}}.}
}
\value{
A \href{https://patchwork.data-imaginist.com/index.html}{patchwork}
of ggplots.
}
\description{
Plots a comparison of the samples joint component scores obtained for the two
datasets in an sO2PLS run (from the \code{OmicsPLS} package).
}
