% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_crossval_o2m}
\alias{so2pls_crossval_o2m}
\title{Wrapper for OmicsPLS::crossval_o2m function}
\usage{
so2pls_crossval_o2m(
  omicspls_input,
  cv_adj_res = NULL,
  a = 1:5,
  ax = seq(0, 10, by = 2),
  ay = seq(0, 10, by = 2),
  nr_folds = 10,
  ...
)
}
\arguments{
\item{omicspls_input}{A named list of length 2, produced by \code{\link{get_input_omicspls}}.}

\item{cv_adj_res}{Data-frame returned by \code{\link[OmicsPLS]{crossval_o2m_adjR2}}
or \code{\link{so2pls_crossval_o2m_adjR2}}. Default value is \code{NULL}.}

\item{a}{Vector of positive integers, number of joint components to test.
Ignored if \code{cv_adj_res} is not \code{NULL}.}

\item{ax}{Vector of non-negative integers, number of specific components to test
for the first dataset. Ignored if \code{cv_adj_res} is not \code{NULL}.}

\item{ay}{Vector of non-negative integers, number of specific components to test
for the second dataset. Ignored if \code{cv_adj_res} is not \code{NULL}.}

\item{nr_folds}{Positive integer, number of folds to use for the cross-validation.
Default value is \code{10}.}

\item{...}{Further arguments passed to the \code{\link[OmicsPLS]{crossval_o2m_adjR2}} function.}
}
\value{
A list of class "cvo2m" with the original and sorted Prediction
errors and the number of folds used.
}
\description{
Wrapper function around the \code{\link[OmicsPLS]{crossval_o2m}} function.
The main purpose of this wrapper is to add to the result the names of the datasets
to facilitate plotting. If the result of a previous call to \code{\link[OmicsPLS]{crossval_o2m_adjR2}}
or \code{\link{so2pls_crossval_o2m_adjR2}} is provided, will be used
to set values to test for \code{a}, \code{ax} and \code{ay}.
}
\details{
If the result of a previous call to \code{\link[OmicsPLS]{crossval_o2m_adjR2}}
or \code{\link{so2pls_crossval_o2m_adjR2}} is provided through the
\code{cv_adj_res} parameter, the optimal values for \code{n}, \code{nx} and \code{ny} are
extracted from it, and the values of \code{a}, \code{ax} and \code{ay} are set as follows:
\itemize{
\item \code{a} = max(\code{n} - 1, 1):(\code{n} + 1)
\item \code{ax} = max(\code{nx} - 1, 0):(\code{nx} + 1)
\item \code{ay} = max(\code{ny} - 1, 0):(\code{ny} + 1)
}
}
