% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_crossval_o2m_adjR2}
\alias{so2pls_crossval_o2m_adjR2}
\title{Wrapper for OmicsPLS::crossval_o2m_adjR2 function}
\usage{
so2pls_crossval_o2m_adjR2(
  omicspls_input,
  a = 1:5,
  ax = seq(0, 10, by = 2),
  ay = seq(0, 10, by = 2),
  nr_folds = 10,
  ...
)
}
\arguments{
\item{omicspls_input}{A named list of length 2, produced by \code{\link{get_input_omicspls}}.}

\item{a}{Vector of positive integers, number of joint components to test.}

\item{ax}{Vector of non-negative integers, number of specific components to test
for the first dataset.}

\item{ay}{Vector of non-negative integers, number of specific components to test
for the second dataset.}

\item{nr_folds}{Positive integer, number of folds to use for the cross-validation.
Default value is \code{10}.}

\item{...}{Further arguments passed to the \code{\link[OmicsPLS]{crossval_o2m_adjR2}} function.}
}
\value{
A data-frame with four columns: \code{MSE}, \code{n}, \code{nx} and \code{ny.} Each row
corresponds to an element in \code{a}.
}
\description{
Wrapper function around the \code{\link[OmicsPLS]{crossval_o2m_adjR2}} function.
The main purpose of this wrapper is to add to the result the names of the datasets
to facilitate plotting.
}
