% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_get_components}
\alias{so2pls_get_components}
\title{Get list of latent components from sO2PLS results}
\usage{
so2pls_get_components(so2pls_res)
}
\arguments{
\item{so2pls_res}{The sO2PLS results generated with \code{\link[=get_output_so2pls]{get_output_so2pls()}}
function.}
}
\value{
A list with the following elements:
\itemize{
\item \code{joint}: character vector with the name of the joint latent components.
\item \code{specific}: named list of length 2. Each element corresponds to a dataset
(names of the list are the datasets name), and is a character vector with
the name of the specific latent components for the corresponding dataset.
}
}
\description{
Extracts the list of joint and specific latent components sO2PLS results.
}
