% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_get_optim_ncomp}
\alias{so2pls_get_optim_ncomp}
\title{Extract optimal number of components from cross-validation results for sO2PLS}
\usage{
so2pls_get_optim_ncomp(cv_res)
}
\arguments{
\item{cv_res}{A \code{cvo2m} object, output from the \code{\link[OmicsPLS]{crossval_o2m}} function.}
}
\value{
A vector with three integer values:
\itemize{
\item \code{n}: optimal number of joint components
\item \code{nx}: optimal number of specific components for dataset X (first dataset)
\item \code{ny}: optimal number of specific components for dataset Y (second dataset)
}
}
\description{
Extracts the optimal number of components (joint and dataset-specific) estimated
via cross-validation results for sO2PLS.
}
