% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_get_variance_explained}
\alias{so2pls_get_variance_explained}
\title{Percentage of variance explained for sO2PLS}
\usage{
so2pls_get_variance_explained(so2pls_res)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}
}
\value{
A tibble with columns \code{latent_dimension}, \code{dataset} and \code{prop_var_expl}.
}
\description{
Generates a table giving the percentage of variance explained by each
component from an sO2PLS in the corresponding dataset.
}
