% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_get_wa_coord}
\alias{so2pls_get_wa_coord}
\title{Computes average sample coordinates for sO2PLS joint components}
\usage{
so2pls_get_wa_coord(so2pls_res)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}
}
\value{
A matrix of samples coordinates, with samples in rows and joint
components in columns.
}
\description{
Computes the average sample coordinates for sO2PLS joint components
across the two datasets.
}
