% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_o2m}
\alias{so2pls_o2m}
\title{Wrapper for OmicsPLS::o2m function}
\usage{
so2pls_o2m(
  omicspls_input,
  cv_res = NULL,
  sparsity_res = NULL,
  n = NULL,
  nx = NULL,
  ny = NULL,
  sparse = FALSE,
  keepx = NULL,
  keepy = NULL,
  ...
)
}
\arguments{
\item{omicspls_input}{A named list of length 2, produced by \code{\link{get_input_omicspls}}.}

\item{cv_res}{Named integer vector of length 3, with names \code{n}, \code{nx}, \code{ny}.
Should be obtained with \code{\link{so2pls_get_optim_ncomp_adj}} or
\code{\link{so2pls_get_optim_ncomp}}.}

\item{sparsity_res}{Named list of length 2, with names \code{keepx} and \code{keepy}.
Should be obtained with \code{\link{so2pls_get_optim_keep}}.}

\item{n}{Positive integer, number of joint components to compute.
Ignored if \code{cv_res} is not \code{NULL}.}

\item{nx}{Positive integer, number of specific components to compute for the
first dataset. Ignored if \code{cv_res} is not \code{NULL}.}

\item{ny}{Positive integer, number of specific components to compute for the
second dataset. Ignored if \code{cv_res} is not \code{NULL}.}

\item{sparse}{Logical, should feature selection be performed? Default value
is \code{FALSE}. If \code{sparsity_res} is not \code{NULL}, will be set to \code{TRUE}.}

\item{keepx}{Integer or integer vector of length \code{n}, number of features from the first
dataset to retain for each joint component. Ignored if \code{sparsity_res} is not \code{NULL}.}

\item{keepy}{Integer or integer vector of length \code{n}, number of features from the second
dataset to retain for each joint component. Ignored if \code{sparsity_res} is not \code{NULL}.}

\item{...}{Other arguments passed to \code{\link[OmicsPLS]{o2m}}.}
}
\value{
A list (see \code{\link[OmicsPLS]{o2m}}).
}
\description{
Wrapper function around the \code{\link[OmicsPLS]{o2m}} function.
The main purpose of this wrapper is to add to the result the names of the datasets
to facilitate plotting.
}
