% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_plot_cv}
\alias{so2pls_plot_cv}
\title{Plots cross-validation results for sO2PLS}
\usage{
so2pls_plot_cv(cv_res, nb_col = NULL)
}
\arguments{
\item{cv_res}{A \code{cvo2m} object, output from the \code{\link[OmicsPLS]{crossval_o2m}} function.}

\item{nb_col}{Integer, the number of columns to use for the faceted plot. Default value is
\code{NULL} (the number of columns will be chosen automatically).}
}
\value{
A ggplot.
}
\description{
Plots the results of a cross-validation for an sO2PLS
run (from the \code{OmicsPLS} package)
}
