% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_plot_cv_adj}
\alias{so2pls_plot_cv_adj}
\title{Plot adjusted cross-validation results for sO2PLS}
\usage{
so2pls_plot_cv_adj(cv_res, with_labels = TRUE)
}
\arguments{
\item{cv_res}{Data-frame, output from the \code{\link[OmicsPLS]{crossval_o2m_adjR2}} function.}

\item{with_labels}{Boolean, whether the optimal values for \code{nx} and \code{ny} for each value
of \code{n} should be displayed. Default value is \code{TRUE}.}
}
\value{
A ggplot
}
\description{
Plots the results of an adjusted cross-validation for an sO2PLS
run (from the \code{OmicsPLS} package)
}
