% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_plot_joint_components_coefficients}
\alias{so2pls_plot_joint_components_coefficients}
\title{Plots sO2PLS contributions between datasets joint components}
\usage{
so2pls_plot_joint_components_coefficients(so2pls_res, datasets = NULL)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}

\item{datasets}{Optional, a character vector with the names of the datasets
that should be plotted. Default is \code{NULL}, i.e. both datasets are considered.}
}
\value{
A \href{https://patchwork.data-imaginist.com/index.html}{patchwork} of ggplots.
}
\description{
Plots the regression coefficients that link the joint components of the two
datasets, from an SO2PLS run (from the \code{OmicsPLS} package).
}
