% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_plot_samples_joint_components}
\alias{so2pls_plot_samples_joint_components}
\title{Plots sO2PLS joint components samples scores}
\usage{
so2pls_plot_samples_joint_components(so2pls_res, ...)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}

\item{...}{Further arguments passed to \code{\link[=plot_samples_score]{plot_samples_score()}}.}
}
\value{
A ggmatrix plot.
}
\description{
Plots the samples scores for the average joint components from an
sO2PLS run (from the \code{OmicsPLS} package).
}
