% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_plot_summary}
\alias{so2pls_plot_summary}
\title{Plot summary of sO2PLS run}
\usage{
so2pls_plot_summary(so2pls_res, datasets = NULL)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}

\item{datasets}{Optional, a character vector with the names of the datasets
for which selected features should be extracted. Default is \code{NULL}, i.e. both
datasets are considered.}
}
\value{
A \href{https://patchwork.data-imaginist.com/index.html}{patchwork} of ggplots.
}
\description{
Plots a summary of variation from an sO2PLS run (from
the \code{OmicsPLS} package).
}
