% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_print_cv_adj}
\alias{so2pls_print_cv_adj}
\title{Print adjusted cross-validation results for sO2PLS}
\usage{
so2pls_print_cv_adj(cv_res)
}
\arguments{
\item{cv_res}{Data-frame, output from the \code{\link[OmicsPLS]{crossval_o2m_adjR2}} function.}
}
\value{
A tibble.
}
\description{
Prints the results of an adjusted cross-validation for an sO2PLS
run (from the \code{OmicsPLS} package)
}
