% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_print_cv_sparsity}
\alias{so2pls_print_cv_sparsity}
\title{Print sparsity cross-validation results for sO2PLS}
\usage{
so2pls_print_cv_sparsity(cv_res_optim)
}
\arguments{
\item{cv_res_optim}{Named list, output from the \code{\link{so2pls_get_optim_keep}} function.}
}
\value{
A tibble, giving for each dataset (\code{dataset} column) and joint component
(other columns) the optimal number of features to retain, as well as the total number
of features per dataset to retain.
}
\description{
Prints the results of a sparsity cross-validation for an sO2PLS
run (from the \code{OmicsPLS} package)
}
