% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2pls.R
\name{so2pls_screeplot}
\alias{so2pls_screeplot}
\title{Screeplot sO2PLS run}
\usage{
so2pls_screeplot(so2pls_res, datasets = NULL)
}
\arguments{
\item{so2pls_res}{The output from the \code{\link[OmicsPLS]{o2m}} function.}

\item{datasets}{Optional, a character vector with the names of the datasets
for which selected features should be extracted. Default is \code{NULL}, i.e. both
datasets are considered.}
}
\value{
A \href{https://patchwork.data-imaginist.com/index.html}{patchwork} of ggplots.
}
\description{
Plots the percentage of variation explained by each latent component in
an sO2PLS (from the \code{OmicsPLS} package).
}
\details{
Note that the plots are set up so that it is possible to add a custom colour
palette to get different colours for each dataset (see example).
}
\examples{
\dontrun{
## by default, same colour used for both datasets (cannot find a way to fix that cleanly)
so2pls_screeplot(so2pls_final_res)

## Add a colour palette to get different colour for each dataset
so2pls_screeplot(so2pls_final_res) &
 scale_fill_brewer(palette = "Set1", drop = F)
}
}
