% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spls.R
\name{spls_get_optim_ncomp}
\alias{spls_get_optim_ncomp}
\title{Select the optimal ncomp from sPLS cross-validation results}
\usage{
spls_get_optim_ncomp(spls_perf, thr = 0.0975, min_ncomp = 1)
}
\arguments{
\item{spls_perf}{List, the result of \code{\link[mixOmics:perf]{mixOmics::perf()}}.}

\item{thr}{Numeric, the threshold to be used for the \code{Q2} values. Default
value is 0.0975.}

\item{min_ncomp}{Integer, the minimum \code{ncomp} value to be returned. Default
value is 1, i.e. this argument does not play a role in selecting the \code{comp}
value. Can be useful if we want at least 2 latent components for final
plots.}
}
\value{
An integer, the optimal number of components to use for the sPLS
run.
}
\description{
Select the optimal number of components to compute in a sPLS run from the
cross-validation results obtained with \code{\link[mixOmics:perf]{mixOmics::perf()}} on a PLS or sPLS
result, using the the mean \code{Q2.total} values. Note that this function is
experimental, the corresponding diagnostic plots should be considered when
selecting the optimal number of components to use.
}
\details{
The selection is made as follows:
\itemize{
\item If all \code{Q2} values are below the threshold specified with \code{thr}, the
number of components yielding the highest \code{Q2} value is selected.
\item If all \code{Q2} values are above the threshold, the number of components
yielding the lowest \code{Q2} value is selected.
\item If the \code{Q2} values are increasing, the number of components \code{n} is
selected such that \code{n+1} is the smallest number of components with a \code{Q2}
value above the threshold.
\item If the \code{Q2} values are decreasing, the number of components \code{n} is
selected such that \code{n+1} is the smallest number of components with a \code{Q2}
value below the threshold.
}
}
