% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spls.R
\name{spls_get_params}
\alias{spls_get_params}
\title{Get parameters from sPLS run}
\usage{
spls_get_params(spls_res)
}
\arguments{
\item{spls_res}{The output from \code{\link[mixOmics]{spls}} or
\code{\link{spls_run}}.}
}
\value{
A tibble.
}
\description{
Extracts the \code{ncomp}, \code{keepX} and \code{keepY} parameters from a sPLS run and
format them into a table.
}
