% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spls.R
\name{spls_get_wa_coord}
\alias{spls_get_wa_coord}
\title{Computes average sample coordinates for sPLS components}
\usage{
spls_get_wa_coord(spls_res)
}
\arguments{
\item{spls_res}{The output from the \code{\link[=spls_run]{spls_run()}} or \code{\link[mixOmics:spls]{mixOmics::spls()}}
function.}
}
\value{
A matrix of samples coordinates, with samples in rows and joint
components in columns.
}
\description{
Computes the average sample coordinates for sPLS components across the two
datasets.
}
