% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_bestNormalise_auto}
\alias{transform_bestNormalise_auto}
\title{Applies the bestNormalize function to rows of a matrix}
\usage{
transform_bestNormalise_auto(mat, return_matrix_only = FALSE, ...)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{return_matrix_only}{Logical, should only the transformed matrix be returned? If \code{TRUE},
the function will return a matrix. If \code{FALSE}, the function instead returns a list with the
transformed data and potentially other information relevant to the transformation. Default
value is \code{FALSE}.}

\item{...}{Further arguments passed to the \code{\link[bestNormalize]{bestNormalize}} function.}
}
\value{
Depending on the \code{return_matrix_only}, either a matrix of transformed data, or
a list with the following elements:
\itemize{
\item \code{transformed_data}: matrix of the transformed data;
\item \code{info_transformation}: A named list with one element per feature (row), giving
details of the transformation applied to the feature (see output of \code{\link[bestNormalize]{bestNormalize}}).
}
}
\description{
Applies an appropriate normalisation method to each feature (row) in a matrix,
via the \code{\link[bestNormalize]{bestNormalize}} function from the \code{bestNormalize}
package.
}
