% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_bestNormalise_manual}
\alias{transform_bestNormalise_manual}
\title{Applies a normalisation method from bestNormalize to rows of a matrix}
\usage{
transform_bestNormalise_manual(mat, method, return_matrix_only = FALSE, ...)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{method}{Character, name of the normalisation method to apply. Possible values are
\code{"arcsinh_x"}, \code{"boxcox"}, \code{"center_scale"}, \code{"exp_x"}, \code{"log_x"}, \code{"orderNorm"},
\code{"sqrt_x"}, \code{"yeojohnson"}. See Details.}

\item{return_matrix_only}{Logical, should only the transformed matrix be returned? If \code{TRUE},
the function will return a matrix. If \code{FALSE}, the function instead returns a list with the
transformed data and potentially other information relevant to the transformation. Default
value is \code{FALSE}.}

\item{...}{Further arguments passed to the \code{method} function from the \code{bestNormalize} package.}
}
\value{
Depending on the \code{return_matrix_only}, either a matrix of transformed data, or
a list with the following elements:
\itemize{
\item \code{transformed_data}: matrix of the transformed data;
\item \code{info_transformation}: A named list with one element per feature (row), giving
details of the transformation applied to the feature (see output for the \code{bestNormalize}
function corresponding to \code{method}).
}
}
\description{
Applies an chosen normalisation method to each feature (row) in a matrix,
via the \code{bestNormalize} package.
}
\details{
Applies a normalisation method implemented in the \code{bestNormalize} package.
The \code{method} argument corresponds to the function from the \code{bestNormalize}
package that will be applied to the rows of the matrix. See the
\href{https://cran.r-project.org/web/packages/bestNormalize/vignettes/bestNormalize.html}{vignette from the \code{bestNormalize} package}
for more information about the transformations.
}
