% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_logx}
\alias{transform_logx}
\title{Applies a log-x transformation to matrix}
\usage{
transform_logx(
  mat,
  return_matrix_only = FALSE,
  log_base = 2,
  pre_log_function = zero_to_half_min
)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{return_matrix_only}{Logical, should only the transformed matrix be
returned? If \code{TRUE}, the function will return a matrix. If \code{FALSE}, the
function instead returns a list with the transformed data and potentially
other information relevant to the transformation. Default value is \code{FALSE}.}

\item{log_base}{Numeric, the base with respect to which logarithms are
computed.}

\item{pre_log_function}{Function that will be applied to the matrix before
the log transformation (e.g. to apply an offset to the values to avoid
issues with zeros). Default value is the \code{\link[=zero_to_half_min]{zero_to_half_min()}} function.}
}
\value{
Depending on the \code{return_matrix_only}, either a matrix of
transformed data, or a list with the following elements:
\itemize{
\item \code{transformed_data}: matrix of the transformed data;
\item \code{info_transformation}: a list with the log base used and the function
applied prior to log-transformation.
}
}
\description{
Applies a log-x transformation (by default log2) through the \code{\link[=log]{log()}}
function.
}
