% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_vsn}
\alias{transform_vsn}
\title{Applies Variance Stabilising Normalisation (vsn) to matrix}
\usage{
transform_vsn(mat, return_matrix_only = FALSE, ...)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{return_matrix_only}{Logical, should only the transformed matrix be returned? If \code{TRUE},
the function will return a matrix. If \code{FALSE}, the function instead returns a list with the
transformed data and potentially other information relevant to the transformation. Default
value is \code{FALSE}.}

\item{...}{Further arguments passed to \code{\link[vsn]{vsn2}}.}
}
\value{
Depending on the \code{return_matrix_only}, either a matrix of transformed data, or
a list with the following elements:
\itemize{
\item \code{transformed_data}: matrix of the transformed data;
\item \code{info_transformation}: NULL.
}
}
\description{
Applies the Variance Stabilising Normalisation performed by the \code{vsn} package
via the \code{\link[vsn]{justvsn}} function.
}
