% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_vst}
\alias{transform_vst}
\title{Applies Variance Stabilising Transformation (DESeq2) to matrix}
\usage{
transform_vst(mat, return_matrix_only = FALSE)
}
\arguments{
\item{mat}{Numeric matrix, must contain integers only.}

\item{return_matrix_only}{Logical, should only the transformed matrix be returned? If \code{TRUE},
the function will return a matrix. If \code{FALSE}, the function instead returns a list with the
transformed data and potentially other information relevant to the transformation. Default
value is \code{FALSE}.}
}
\value{
Depending on the \code{return_matrix_only}, either a matrix of transformed data, or
a list with the following elements:
\itemize{
\item \code{transformed_data}: matrix of the transformed data;
\item \code{info_transformation}: A \code{\link[DESeq2]{DESeqTransform}} object, with details about the transformation.
}
}
\description{
Applies the Variance Stabilising Transformation (VST) performed by the \code{DESeq2} package
via the \code{\link[DESeq2]{varianceStabilizingTransformation}} function. Includes a
size factor normalisation prior to the VST. Only applies to a matrix of count.
}
