% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{zero_to_half_min}
\alias{zero_to_half_min}
\title{Replace zeros with half-min in matrix}
\usage{
zero_to_half_min(mat)
}
\arguments{
\item{mat}{Numeric matrix.}
}
\value{
The matrix with zero values replaced.
}
\description{
Replace zero values in a matrix by half of the minimum non-null value in the
matrix.
}
