% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_omics_sets.R
\name{create_multiomics_set}
\alias{create_multiomics_set}
\title{Create a MultiDataSet object to store multi-omics data}
\usage{
create_multiomics_set(sets_list, datasets_names = NULL, show_warnings = TRUE)
}
\arguments{
\item{sets_list}{List of Biobase Set objects, created via \code{\link{create_omics_set}}. Currently accepted ojects: \code{SnpSet},
\code{ExpressionSet}, \code{MetabolomeSet}, \code{PhenotypeSet}.}

\item{datasets_names}{Optional, vector of character, name for each Set object. Will be appended to the data type in the resulting
object. If the \code{sets_list} list contains several objects of the same data type (e.g. several SnpSets), their names must be unique.
If "" is provided, no name will be appended to the data type for the corresponding dataset.}

\item{show_warnings}{Logical, should warnings be displayed when adding a set to the MultiDataSet object? Default value is \code{TRUE}.}
}
\value{
a \code{\link[MultiDataSet]{MultiDataSet-class}}.
}
\description{
Creates a MultiDataSet object from a list of Biobase Set objects to store the different omics sets.
}
