% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_comparison.R
\name{.heatmap_features_weight_corr}
\alias{.heatmap_features_weight_corr}
\title{Heatmap of features weight correlation}
\usage{
.heatmap_features_weight_corr(
  output_list,
  include_missing_features = FALSE,
  legend_ncol = 1
)
}
\arguments{
\item{output_list}{List of integration methods output generated via the
\code{get_output()} function.}

\item{include_missing_features}{Logical, see
\code{\link[=get_features_weight_correlation]{get_features_weight_correlation()}} for details. Default value is \code{FALSE}.}

\item{legend_ncol}{Integer, number of columns in the legend. Default value is
\code{1}.}
}
\value{
a \code{ComplexHeatmap::Heatmap} (lower triangle only).
}
\description{
Constructs a lower triangle heatmap of features weight correlation between
the latent dimensions constructed by several integration methods.
}
