% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidataset_utils.R
\name{n_features}
\alias{n_features}
\title{Number of features in each dataset of MultiDataSet object}
\usage{
n_features(mo_data)
}
\arguments{
\item{mo_data}{A \code{\link[MultiDataSet]{MultiDataSet-class}} object.}
}
\value{
A named vector, where each element is the number of features in the
corresponding dataset.
}
\description{
Gives the number of features in each dataset of a MultiDataSet object.
}
